% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{about_variable}
\alias{about_variable}
\title{about_variable}
\usage{
about_variable(variable, model_data)
}
\arguments{
\item{variable}{A variable in an mcmc object output by \code{\link{fit_model}}}

\item{model_data}{The model training data created by \code{\link{format_model_data}} and input to \code{\link{fit_model}}.}
}
\value{
Text that explains the variable
}
\description{
\code{about_variable()} returns information about the model variable.
}
\examples{
about_variable(variable = "mu[1,2]", model_data = example_model_data)
about_variable(variable = "gamma[5]", model_data = example_model_data)

}
\keyword{model}
