% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{add_character_features}
\alias{add_character_features}
\title{add_character_features}
\usage{
add_character_features(img, letterList, letters, dims)
}
\arguments{
\item{img}{thinned binary image}

\item{letterList}{list containing letter characters}

\item{letters}{individual characters from letterList}

\item{dims}{image graph dimensions}
}
\value{
a list of letters with features applied
}
\description{
Internal method that adds features to characters
}
