% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{add_covariance_matrix}
\alias{add_covariance_matrix}
\title{add_covariance_matrix}
\usage{
add_covariance_matrix(character_lists, character_features, dims)
}
\arguments{
\item{character_lists}{Output from processHandwriting$letterLists}

\item{character_features}{Nested lists associating features to respective characters.}

\item{dims}{Dimensions of binary image}
}
\value{
nested lists associating features to respective characters.
}
\description{
add_covariance_matrix
}
\keyword{centroid,}
\keyword{character}
\keyword{lean,}
\keyword{skew,}
\keyword{slant,}
