% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{add_line_info}
\alias{add_line_info}
\title{add_line_info}
\usage{
add_line_info(character_features, img_dim)
}
\arguments{
\item{character_features}{All extracted features}

\item{img_dim}{Dimensions of binary image}
}
\value{
Appends line information to character features
}
\description{
Associates characters to their respective line numbers
Needs improvement if runtime becomes a problem
}
\keyword{character,}
\keyword{features,}
\keyword{line}
\keyword{number}
