% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{AddLetterImages}
\alias{AddLetterImages}
\title{Add Letter Images}
\usage{
AddLetterImages(letterList, dims)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{dims}{Dimensions of the original document}
}
\value{
letterList with a new matrix \code{image} value for each sublist.
}
\description{
Pulls out letterlist as its own object, and adds the image matrix as well
}
\examples{
twoSent_document = list()
twoSent_document$image = twoSent
twoSent_document$thin = thinImage(twoSent_document$image)
twoSent_processList = processHandwriting(twoSent_document$thin, dim(twoSent_document$image))

dims = dim(twoSent_document$image)
withLetterImages = AddLetterImages(twoSent_processList$letterList, dims)

}
