% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{AddSamplingStrata}
\alias{AddSamplingStrata}
\title{AddSamplingStrata}
\usage{
AddSamplingStrata(letterList)
}
\arguments{
\item{letterList}{List of letters}
}
\value{
List of letters
}
\description{
Add two new fields, numloops and stratum, to the characterFeatures list for
each letter in letterList.
}
\keyword{?}
