% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addToFeatures}
\alias{addToFeatures}
\title{addToFeatures}
\usage{
addToFeatures(FeatureSet, LetterList, vectorDims)
}
\arguments{
\item{FeatureSet}{The current list of features that have been calculated}

\item{LetterList}{List of all letters and their information}

\item{vectorDims}{Vectors with image Dims}
}
\value{
A list consisting of current features calculated in FeatureSet as well as measures of compactness, loop count, and loop dimensions
}
\description{
addToFeatures
}
