% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{all_centroids}
\alias{all_centroids}
\title{all_centroids}
\usage{
all_centroids(character_features)
}
\arguments{
\item{character_features}{Features extracted from any given document}
}
\value{
All centroids concatenated with one another (unlisted)
}
\description{
Iterates through extracted character features, extracting
all centroids found for later use in line numbering
}
\keyword{character,}
\keyword{line}
\keyword{loops,}
