% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{all_down_dists}
\alias{all_down_dists}
\title{all_down_dists}
\usage{
all_down_dists(character_features)
}
\arguments{
\item{character_features}{Features extracted from any given document}
}
\value{
All downdistance concatenated with one another (unlisted)
}
\description{
Iterates through extracted character features, extracting
all downward distances found for later use in line separating
}
\keyword{character,}
\keyword{line}
\keyword{neighbor,}
