% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{AllUniquePaths}
\alias{AllUniquePaths}
\title{Internal function for getting a list of all non loop paths in a writing sample.}
\usage{
AllUniquePaths(adj, graph, graph0)
}
\description{
Internal function for getting a list of all non loop paths in a writing sample.
}
