% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{analyze_questioned_documents2}
\alias{analyze_questioned_documents2}
\title{analyze_questioned_documents2}
\usage{
analyze_questioned_documents2(
  template_dir,
  questioned_images_dir,
  model,
  num_cores,
  writer_indices,
  doc_indices
)
}
\arguments{
\item{template_dir}{A directory that contains a cluster template created by \code{\link{make_clustering_templates}}}

\item{questioned_images_dir}{A directory containing questioned documents}

\item{model}{A fitted model created by \code{\link{fit_model}}}

\item{num_cores}{An integer number of cores to use for parallel processing
with the \code{doParallel} package.}

\item{writer_indices}{A vector of start and stop characters for writer IDs in file names}

\item{doc_indices}{A vector of start and stop characters for document names in file names}
}
\value{
A list of likelihoods, votes, and posterior probabilities of
writership for each questioned document.
}
\description{
\code{analyze_questioned_documents2()} estimates the posterior probability of
writership for the questioned documents using MCMC draws from a hierarchical
model created with \code{fit_model()}.
}
\examples{
\dontrun{
template_dir <- "/path/to/template"
questioned_images_dir <- "/path/to/questioned_images"
analysis <- analyze_questioned_documents2(
  template_dir = template_dir
  questioned_images_dir = questioned_images_dir
  model = model,
  num_cores = 2,
  writer_indices = c(2,5),
  doc_indices = c(7,18)
)
analysis$posterior_probabilities
}

}
\keyword{model}
