% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{calculate_accuracy}
\usage{
calculate_accuracy(analysis)
}
\arguments{
\item{analysis}{Writership analysis output by
\code{\link{analyze_questioned_documents}}}
}
\value{
The model's accuracy on the test set
}
\description{
\code{calculate_accuracy} measures the accuracy of fitted model on a test set of
documents by calculating the average posterior probability assigned to the
true writer. Fit a model with \code{\link{fit_model}} and calculate posterior
probabilities of writership with \code{\link{analyze_questioned_documents}}.
}
