% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{centeredImage}
\alias{centeredImage}
\title{centeredImage}
\usage{
centeredImage(letter)
}
\arguments{
\item{letter}{A letter}
}
\value{
a named list with fields nodes, centroid, proptroid, image and allPaths
}
\description{
Find the letter's centroid and proptroid relative to the bottom left corner of the
letter's image.
}
