% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{centeredImageOnCentroid}
\alias{centeredImageOnCentroid}
\title{centeredImageOnCentroid}
\usage{
centeredImageOnCentroid(letter)
}
\arguments{
\item{letter}{A letter from a handwriting sample}
}
\value{
The letter with it's paths and path ends locations given relative to the 
letter's centroid.
}
\description{
Place the letter's centroid at (0,0) and find the locations (column and row
numbers) of a letter's paths and path ends relative to the letter's centroid.
}
