% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{checkStacking}
\alias{checkStacking}
\title{checkStacking}
\usage{
checkStacking(candidateBreaks, allPaths, letters, nodeGraph0, dims)
}
\arguments{
\item{candidateBreaks}{possible breaks for letterpath}

\item{allPaths}{list of paths}

\item{letters}{list of individual letter characters}

\item{nodeGraph0}{skeletonized graph}

\item{dims}{graph dimensions}
}
\value{
stackPtFlag
}
\description{
Internal function for removing breakpoints that follow all of the rules, but separate two letters that are stacked on top of each other.
}
