% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{cleanBinaryImage}
\alias{cleanBinaryImage}
\title{cleanBinaryImage}
\usage{
cleanBinaryImage(img, clean = TRUE, fill = TRUE, cleanAmt = 2,
  fillAmt = 2)
}
\arguments{
\item{img}{Binary image to clean.}

\item{clean}{Boolean for whether or not to shrink then grow image.}

\item{fill}{Boolean for whether or not to grow then shrink image.}

\item{amount}{Parameter passed to clean and fill. Strength of clean and/or fill.}
}
\value{
Cleaned binary image.
}
\description{
Function to clean up a binary image. If clean is true, then will shrink by 1 pixel and then regrow by 1 pixel.
If fill is true, then will grow by 1 pixel then shrink by 1 pixel. By default it does both - shrinks then grows.
}
