% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{coordsToIndex}
\alias{coordsToIndex}
\title{plotImage}
\usage{
coordsToIndex(rowVals, colVals, dimImageAt1)
}
\arguments{
\item{x}{Binary matrix, usually from readPNGBinary}
}
\value{
Returns plot of x.
}
\description{
This function plots a basic binary image.
}
\keyword{plot}
