% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCSAFEDocs.R
\name{copy_csafe_docs}
\alias{copy_csafe_docs}
\title{copy_CSAFE_docs}
\usage{
copy_csafe_docs(docs, input_dir, template_dir, model_dir, questioned_dir)
}
\arguments{
\item{docs}{A list of CSAFE documents created by \code{\link{select_csafe_docs}}}

\item{input_dir}{Path to directory containing the downloaded CSAFE
Handwriting Database}

\item{template_dir}{Path to directory where template training documents will
be copied}

\item{model_dir}{Path to directory where model training documents will be
copied}

\item{questioned_dir}{Path to}
}
\value{
Message "Documents have been copied."
}
\description{
The user can build 3 sets - template training, model training, and questioned
(test) - of documents from the CSAFE Handwriting Database with the function
\code{\link{select_csafe_docs}}. The output of \code{select_csafe_docs} is a list of three
data frames - template, model, and questioned. Each data frame contains the
file names of CSAFE documents assigned to either the template, model, or
questioned set. The \code{copy_csafe_docs} function will copy these sets of
documents from the directory containing the downloaded CSAFE Handwriting
Database to the new directories.
}
\examples{
\dontrun{
docs <- select_csafe_docs(
  num_template_writers = 10,
  template_sessions = 1,
  template_reps = 1,
  template_prompts = "London Letter",
  template_seed = 100,
  num_model_writers = 5,
  model_sessions = 1,
  model_reps = c(1, 2, 3),
  model_prompts = "Wizard of Oz",
  model_seed = 101,
  questioned_sessions = 3,
  questioned_reps = 1,
  questioned_prompts = "Wizard of Oz"
)

input_dir <- path / to / downloaded_CSAFE_docs
output_template_dir <- path / to / template_docs_output
output_model_dir <- path / to / model_docs_output
output_questioned_dir <- path / to / questioned_docs_output

copy_csafe_docs(
  docs = docs,
  input_dir = input_dir,
  template_dir = output_template_dir,
  model_dir = output_model_dir,
  questioned_dir = output_questioned_dir
)
}

}
