% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_analysisfunctions.R
\name{count_csafe_correct_top_writer}
\alias{count_csafe_correct_top_writer}
\title{count_csafe_correct_top_writer}
\usage{
count_csafe_correct_top_writer(analysis)
}
\arguments{
\item{analysis}{The output of \code{\link{analyze_questioned_documents}}. If more than
one questioned document was analyzed with this function, then the data
frame analysis$posterior_probabilities lists the posterior probabilities
for all questioned documents. \code{show_posterior_probabilities} creates a data
frame of the posterior probabilities for a single questioned document and
sorts the known writers from the most likely to least likely to have writen
the questioned document.}
}
\value{
A list of the number of correct top writers and the total number of
questioned documents
}
\description{
If a model was created with \code{\link{fit_model}} on a closed-set of CSAFE writers
and test documents from those writers are analyzed with
\code{\link{analyze_questioned_documents}}, \code{count_csafe_correct_top_writer} will count
the number of test documents for which the highest posterior probability is
correctly assigned to the true writer.
}
\examples{
count_csafe_correct_top_writer(analysis = example_analysis_1chain)
count_csafe_correct_top_writer(analysis = example_analysis_2chains)

}
