% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{countChanges}
\alias{countChanges}
\title{Internal function for counting 4-connected components around a pixel.}
\usage{
countChanges(coords, img)
}
\description{
Internal function for counting 4-connected components around a pixel.
}
