% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{countNodes}
\alias{countNodes}
\title{countNodes}
\usage{
countNodes(letterList, nodes)
}
\arguments{
\item{letterList}{list containing letter characters}

\item{nodes}{list of nodes}
}
\value{
number of nodes in letterList
}
\description{
Function for counting nodes in a list of letters.
}
