% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractWordInfo.R
\name{create_words}
\alias{create_words}
\title{create_words}
\usage{
create_words(processList)
}
\arguments{
\item{processList}{Output from processHandwriting - contains all glyph information}
}
\value{
list of word objects
}
\description{
creates word objects based on splits found in processHandwriting
}
