% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{crop}
\alias{crop}
\title{crop}
\usage{
crop(img)
}
\arguments{
\item{img}{Full image matrix to be cropped}
}
\value{
Cropped image matrix.
}
\description{
This function crops an image down so that there is 1 pixel of padding on each side of the outermost 0 points.
}
