% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csafe}
\alias{csafe}
\title{Cursive written word: csafe}
\format{
Binary image matrix. 111 rows and 410 columns.
}
\usage{
csafe
}
\description{
Cursive written word: csafe
}
\examples{
csafe_document <- list()
csafe_document$image <- csafe
plotImage(csafe_document$image)
csafe_document$thin <- thinImage(csafe_document$image)
plotImageThinned(csafe_document$image, csafe_document$thin)
csafe_processList <- processHandwriting(csafe_document$thin, dim(csafe_document$image))
}
\keyword{datasets}
