% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFunctions.R
\name{dist_loc}
\alias{dist_loc}
\title{dist_loc}
\usage{
dist_loc(p1e1, p1e2, p2e1, p2e2)
}
\arguments{
\item{p1e1}{Vector of (x,y) coordinates of the first endpoint of the first
path}

\item{p1e2}{Vector of (x,y) coordinates of the second endpoint of the first
path}

\item{p2e1}{Vector of (x,y) coordinates of the first endpoint of the second
path}

\item{p2e2}{Vector of (x,y) coordinates of the second endpoint of the second
path}
}
\value{
Distances between endpoints
}
\description{
Calculate the distances between endpoints of one path with the endpoints of a
second path. Label the endpoints of the first path p1e1 and p1e2. Label the
endpoints of the second path p2e1 and p2e2. Match endpoints in the plus
direction: find the Euclidean distance between p1e1 and p2e1 and between p1e2
and p2e2. Also match endpoints in the minus direction: find the Euclidean
distance between p1e1 and p2e2 and between p1e2 and p2e1. Return the shortest
distance in the plus direction and the shortest distance in the minus
direction.
}
\keyword{?}
