% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{distXY}
\alias{distXY}
\title{distXY}
\usage{
distXY(xy1, xy2)
}
\arguments{
\item{xy1}{Vector of (x,y) coordinates of the first point}

\item{xy2}{Vector of (x,y) coordinates of the second point}
}
\value{
Euclidean distance between the two points
}
\description{
Calculate the Euclidean distance between two points
}
\keyword{?}
