% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{drop_burnin}
\alias{drop_burnin}
\title{drop_burnin}
\usage{
drop_burnin(model, burn_in)
}
\arguments{
\item{model}{A list of MCMC draws from a model fit with \code{\link[=fit_model]{fit_model()}}.}

\item{burn_in}{An integer number of starting iterations to drop from each MCMC chain.}
}
\value{
A list of data frames of MCMC draws with burn-in dropped.
}
\description{
\code{drop_burnin()} removes the burn-in from the MCMC draws.
}
\examples{
model <- fit_model(
  model_data = example_model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(model = model, burn_in = 250)
analysis <- analyze_questioned_documents(
  model_data = example_model_data,
  model = model,
  questioned_data = example_questioned_data,
  num_cores = 2
)

}
\keyword{model}
