% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_analysis}
\alias{example_analysis}
\title{Example of writership analysis}
\format{
A named list created by \code{\link{analyze_questioned_documents}} with 3 items:
\describe{
\item{likelihoods}{list of data frames where each data frame
contains the likelihoods for a questioned document for each MCMC iteration.}
\item{votes}{list of vote tallys for each questioned document.}
\item{posterior_probabilites}{list of posterior probabilities of writership
for each questioned document and each known writer in the closed set used to train the
hierarchical model.}
}
}
\usage{
example_analysis
}
\description{
Example of writership analysis
}
\examples{
\dontrun{
draws <- fit_model(example_model_data$rjags_data, 4000)
analysis <- analyze_questioned_documents(example_model_data,
  draws,
  example_questioned_data,
  num_cores = 2
)
}
}
\keyword{datasets}
