% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model_1chain}
\alias{example_model_1chain}
\title{Example of a hierarchical model}
\format{
An MCMC list created by \code{\link{fit_model}} with a single chain and 50 MCMC iterations.
The MCMC list contains a single MCMC object with 50 rows and 136 columns. Each row corresponds to
an MCMC iteration and each column corresponds to a variable:
\describe{
\item{eta\link{k}}{Hyper priors for cluster k.}
\item{gamma\link{k}}{The pseudo-cluster count across all writers for cluster k.}
\item{mu\link{w,k}}{The location parameter of a wrapped-Cauchy distribution for writer w and cluster k.}
\item{pi\link{w,k}}{The cluster fill probability for writer w and cluster k.}
\item{tau\link{w,k}}{The scale parameter of a wrapped-Cauchy distribution for writer w and cluster k.}
}
}
\usage{
example_model_1chain
}
\description{
Example of a hierarchical model
}
\examples{
# convert to a data frame and view all variable names
df <- as.data.frame(example_model_1chain[[1]])
colnames(df)

# analyze questioned documents
analysis <- analyze_questioned_documents(
  model_data = example_model_data,
  model = example_model_1chain,
  questioned_data = example_questioned_data,
  num_cores = 2
)
plot_posterior_probabilities(analysis)

}
\keyword{datasets}
