% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model_data}
\alias{example_model_data}
\title{Example of formatted model data}
\format{
A named list created by \code{\link{format_model_data}} with 19 items:
\describe{
\item{Y}{data frame of cluster fill counts for each document}
\item{G}{number of clusters}
\item{D}{number of documents}
\item{W}{number of writers}
\item{docN}{number of graphs in each document}
\item{docwriter}{writer id of each document}
\item{zero_vec}{vector of zeros for used for the zeros-trick}
\item{Gsmall}{number of non-empty clusters}
\item{numletters}{total number of graphs}
\item{pc_wrapped}{principal component rotation angle of each graph}
\item{letterwriter}{writer id of each graph}
\item{lettercluster}{cluster assignment of each graph}
\item{zero_mat}{matrix of zeros for the zeros-trick}
\item{a}{parameter}
\item{b}{parameter}
\item{c}{parameter}
\item{d}{parameter}
\item{e}{paramter}
}
}
\usage{
example_model_data
}
\description{
Example of formatted model data
}
\examples{
model_data <- format_model_data(
  model_proc_list = example_model_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18),
  a = 2, b = 0.25, c = 2, d = 2, e = 0.5
)
model <- fit_model(
  model_data = model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(
  model = model,
  burn_in = 250
)
q_data <- format_questioned_data(
  formatted_model_data = model_data,
  questioned_proc_list = example_questioned_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18)
)
analysis <- analyze_questioned_documents(
  model_data = model_data,
  model = model,
  questioned_data = q_data,
  num_cores = 2
)

}
\keyword{datasets}
