% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_clusters}
\alias{example_questioned_clusters}
\title{Example of list of questioned documents' data for the hierarchical model}
\format{
List created by \code{\link{get_clusterassignment}} with extra fields removed to
make the example file smaller.

List created by \code{\link{get_clusterassignment}} with extra fields removed to
make the example file smaller.
}
\usage{
example_questioned_clusters

example_questioned_clusters
}
\description{
Example of list of questioned documents' data for the hierarchical model

Example of list of questioned documents' data for the hierarchical model
}
\examples{
writer_indices <- c(2, 5)
doc_indices <- c(7, 18)
model_data <- format_questioned_data(
  example_questioned_clusters,
  writer_indices,
  doc_indices
)
}
\keyword{datasets}
