% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_data}
\alias{example_questioned_data}
\title{Example of questioned data formatted for the hierarchical model}
\format{
A named list created by \code{\link{format_questioned_data}} with 2 items:
\describe{
\item{graph_measurements}{data frame that shows the writer ID,
document ID, slope, principal component rotation angle, and wrapped
principal rotation angle for each graph }
\item{cluster_fill_counts}{data frame that shows the number of graphs
assigned to each cluster for each document.}
}
}
\usage{
example_questioned_data
}
\description{
Example of questioned data formatted for the hierarchical model
}
\examples{
model_data <- format_model_data(
  model_proc_list = example_model_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18),
  a = 2, b = 0.25, c = 2, d = 2, e = 0.5
)
model <- fit_model(
  model_data = model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(
  model = model,
  burn_in = 250
)
q_data <- format_questioned_data(
  formatted_model_data = model_data,
  questioned_proc_list = example_questioned_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18)
)
analysis <- analyze_questioned_documents(
  model_data = model_data,
  model = model,
  questioned_data = q_data,
  num_cores = 2
)

}
\keyword{datasets}
