% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_proc_list}
\alias{example_questioned_proc_list}
\title{Example of list of questioned documents' data for the hierarchical model}
\format{
List created by `get_clusterassignment` with extra fields removed to
  make the example file smaller.

List created by `get_clusterassignment` with extra fields removed to
  make the example file smaller.
}
\usage{
example_questioned_proc_list

example_questioned_proc_list
}
\description{
Example of list of questioned documents' data for the hierarchical model

Example of list of questioned documents' data for the hierarchical model
}
\keyword{datasets}
