% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_template}
\alias{example_template}
\title{Example template}
\format{
A list containing a single cluster template created by
\code{\link[=make_clustering_templates]{make_clustering_templates()}}. The cluster template was created by
sorting a stratified sample of 1000 graphs from 10 training documents into
5 clusters with a K-means algorithm. The cluster template is a named list
with 14 items:
\describe{
\item{cluster}{A vector of cluster assignments
for each graph used to create the cluster template.}
\item{centers}{A list of graphs used as the starting cluster centers for the K-means algorithm.}
\item{K}{The number of clusters to build (5) with the K-means algorithm.}
\item{n}{The number of training graphs to use (1000) in the K-means algorithm.}
\item{docnames}{A vector of the file names of the training documents.}
\item{iters}{The maximum number of iterations for the K-means algorithm (3).}
\item{changes}{A vector of the number of graphs that changed clusters on each iteration of the K-means algorithm.}
\item{outlierCutoff}{A vector of the outlier cutoff values calculated on each iteration of the K-means algorithm.}
\item{stop_reason}{The reason the K-means algorithm terminated.}
\item{wcd}{A matrix of the within cluster distances on each iteration of the K-means algorithm. More specifically, the distance between each graph
and the center of the cluster to which it was assigned  on each iteration.}
\item{wcss}{A vector of the within-cluster sum of squares on each iteration of the K-means algorithm.}
\item{rmse}{A vector of the root-mean square error on each iteration of the K-means algorithm.}
\item{DaviesBouldinIndex}{The Davies-Bouldin index on each iteration of the K-means algorithm.}
\item{VarianceRatioCriterion}{The variance-ratio criterion on each iteration of the K-means algorithm.}
}
}
\usage{
example_template
}
\description{
Example template
}
\examples{
\dontrun{
# get cluster assignments for model training documents
model_dir <- "path/to/model_training_docs"
model_proc_list <- get_clusterassignment(example_template, model_dir)
}
}
\keyword{datasets}
