% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGraphs.R
\name{extractGraphs}
\alias{extractGraphs}
\title{Extracts graphs from .png images and saves each by their respective writer}
\usage{
extractGraphs(source_folder = getwd(), save_folder = getwd())
}
\arguments{
\item{source_folder}{path to folder containing .png images}

\item{save_folder}{path to folder where graphs are saved to}
}
\value{
saves graphs in an rds file
}
\description{
extractGraphs
}
\examples{
\dontrun{
sof = "path to folder containing .png images"
saf = "path to folder where graphs will be saved to"
extractGraphs(sof, saf)
}
}
