% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorNodes.R
\name{find_colorpoints}
\alias{find_colorpoints}
\title{find_colorpoints}
\usage{
find_colorpoints(words, dims)
}
\arguments{
\item{words}{List of words and some glyph level information}

\item{dims}{The dimensions of the image (important for r/c features)}
}
\value{
A new list with word level information for each word.
}
\description{
Finds and assigns points for Kneser Triangulation
}
