% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{format_model_data}
\alias{format_model_data}
\title{format_model_data}
\usage{
format_model_data(
  model_proc_list,
  writer_indices,
  doc_indices,
  a = 2,
  b = 0.25,
  c = 2,
  d = 2,
  e = 0.5
)
}
\arguments{
\item{model_proc_list}{List of processed handwriting from a set of documents
created by \code{get_clusterassignment()}. Each item in the list contains the
extracted graphs from a document. These processed documents will be used to train a model.}

\item{writer_indices}{Vector of start and end indices for the writer id in
the document names.}

\item{doc_indices}{Vector of start and end indices for the document id in the
document names.}

\item{a}{Scalar}

\item{b}{Scalar}

\item{c}{Scalar}

\item{d}{Scalar}

\item{e}{Scalar}
}
\value{
List of data formatted for rjags.
}
\description{
\code{format_model_data()} formats the data need for the rjags model.
}
\examples{
model_data <- format_model_data(
  model_proc_list = example_model_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18),
  a = 2, b = 0.25, c = 2, d = 2, e = 0.5
)
model <- fit_model(
  model_data = model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(
  model = model,
  burn_in = 250
)
q_data <- format_questioned_data(
  formatted_model_data = model_data,
  questioned_proc_list = example_questioned_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18)
)
analysis <- analyze_questioned_documents(
  model_data = model_data,
  model = model,
  questioned_data = q_data,
  num_cores = 2
)

}
