% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_centroid_info}
\alias{get_centroid_info}
\title{get_centroid_info}
\usage{
get_centroid_info(character, dims)
}
\arguments{
\item{character}{character to extract information from}

\item{dims}{Dimensions of binary image}
}
\value{
List containing centroid, pixel density,letter 'lean', and all supporting information
}
\description{
Extracts centroid & supporting information from a character
Relevant Features:
Centroid Index: R Index representation of centroid location
Centroid x,y: X,Y representations of the centroid, see ?i_to_rci 
Centroid Horiz Location: How far along horizontally (Represented as a number between 0 and 1) the centroid is in its respective character.
Centroid Vertical Location: How far along vertically (Represented as a number between 0 and 1) the centroid is in its respective character.
Slope: 'Letter Lean', slope found between the centroids of each disjoint half in a single character.
The letter is split in half, each halve's centroid is calculated independently, the slope is taken between the two. 
Box Density: (Dimensions of box around letter width height) / (how much of the document it covers) //Might be a more document as opposed to letter based feature
Pixel Density: Ratio of black to white pixels found in box drawn around the letter.
}
\keyword{centroid}
\keyword{character}
\keyword{lean}
\keyword{skew}
\keyword{slant}
