% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{get_cluster_fill_counts}
\alias{get_cluster_fill_counts}
\title{get_cluster_fill_counts}
\usage{
get_cluster_fill_counts(df)
}
\arguments{
\item{df}{A data frame with columns \code{writer}, \code{doc}, and \code{cluster}. Each
row corresponding to a graph and lists the writer of that graph, the document
from which the graph was obtained, and the cluster to which that graph is assigned.}
}
\value{
A dataframe of cluster fill counts for each document in the input data frame.
}
\description{
\code{get_cluster_fill_counts()} creates a data frame that shows the number of
graphs in each cluster for each input document.
}
\examples{
writer <- c(rep(1, 20), rep(2, 20), rep(3, 20))
docname <- c(rep('doc1',20), rep('doc2', 20), rep('doc3', 20))
doc <- c(rep(1, 20), rep(2, 20), rep(3, 20))
cluster <- sample(3, 60, replace=TRUE)
df <- data.frame(docname, writer, doc, cluster)
get_cluster_fill_counts(df)

}
