% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_clusterassignment.R
\name{get_clusterassignment}
\alias{get_clusterassignment}
\title{get_clusterassignment}
\usage{
get_clusterassignment(
  template_dir,
  input_type,
  num_graphs = "All",
  writer_indices,
  doc_indices,
  num_cores
)
}
\arguments{
\item{template_dir}{Directory containing a cluster template created with \code{make_clustering_templates}}

\item{input_type}{\code{model} or \code{questioned}}

\item{num_graphs}{'All' or integer number of graphs to randomly select from each document.}

\item{writer_indices}{Vector of start and end indices for the writer id in
the document names}

\item{doc_indices}{Vector of start and end indices for the document id in the
document names}

\item{num_cores}{Integer number of cores to use for parallel processing}
}
\value{
list of processed handwriting with cluster assignments for each graph
}
\description{
get_clusterassignment
}
