% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_clusterassignment.R
\name{get_clusters_batch}
\alias{get_clusters_batch}
\title{get_clusters_batch}
\usage{
get_clusters_batch(
  template,
  input_dir,
  output_dir,
  writer_indices,
  doc_indices,
  num_cores
)
}
\arguments{
\item{template}{A cluster template created with \code{\link{make_clustering_templates}}}

\item{input_dir}{A directory containing graphs created with \code{\link{process_batch_dir}}}

\item{output_dir}{Output directory for cluster assignments}

\item{writer_indices}{Vector of start and end indices for the writer id in
the graph file names}

\item{doc_indices}{Vector of start and end indices for the document id in the
graph file names}

\item{num_cores}{Integer number of cores to use for parallel processing}
}
\value{
A list of cluster assignments
}
\description{
get_clusters_batch
}
