% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csafe_db_helpers.R
\name{get_csafe_writerIDs}
\alias{get_csafe_writerIDs}
\title{Extract CSAFE writer IDs}
\usage{
get_csafe_writerIDs(x, only_unique = FALSE, as_integer = FALSE)
}
\arguments{
\item{x}{A character vector of CSAFE document names, file names, or full
file paths, such as "w0001_s01_pLND_r01.png" or
"path/to/w0001_s01_pLND_r01.png".}

\item{only_unique}{Logical. If `TRUE`, return only unique writer IDs.
Defaults to `FALSE`.}

\item{as_integer}{Logical. If `TRUE`, remove the leading "w" and return the
writer IDs as integers. Defaults to `FALSE`.}
}
\value{
A character vector of writer IDs, or an integer vector if
  `as_integer = TRUE`.
}
\description{
Extract writer IDs from CSAFE Handwriting Database document names or file
paths. Writer IDs are the first five characters of each file name (e.g.,
"w0001").
}
\examples{
docs <- c("path/to/w0001_s01_pLND_r01.png", "w0001_s02_pWOZ_r01.png",
          "w0238_s01_pPHR_r02.png")
get_csafe_writerIDs(docs)
get_csafe_writerIDs(docs, only_unique = TRUE)
get_csafe_writerIDs(docs, only_unique = TRUE, as_integer = TRUE)

}
