% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_loop_info}
\alias{get_loop_info}
\title{get_loop_info}
\usage{
get_loop_info(character, img_dim)
}
\arguments{
\item{character}{Target for loop association}

\item{img_dim}{Dimensions of binary image}
}
\value{
Loop information to respective character
}
\description{
Primary driver of loop to character association
}
\keyword{associate}
\keyword{character,}
\keyword{loop,}
