% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{get_posterior_probabilities}
\alias{get_posterior_probabilities}
\title{get_posterior_probabilities}
\usage{
get_posterior_probabilities(analysis, questioned_doc)
}
\arguments{
\item{analysis}{The output of \code{\link{analyze_questioned_documents}}. If more than
one questioned document was analyzed with this function, then the data frame
analysis$posterior_probabilities lists the posterior probabilities for all
questioned documents. \code{show_posterior_probabilities} creates a data frame of the
posterior probabilities for a single questioned document and sorts the known writers
from the most likely to least likely to have writen the questioned document.}

\item{questioned_doc}{The name of a questioned document}
}
\value{
A data frame of posterior probabilities for the questioned document
}
\description{
Get the posterior probabilities for a questioned document.
}
\examples{
show_posterior_probabilities(
  analysis = example_analysis_1chain,
  questioned_doc = "w9_s03_pWOZ_r1."
)
show_posterior_probabilities(
  analysis = example_analysis_1chain,
  questioned_doc = "w30_s03_pWOZ_r1."
)

}
