% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{getGraphDistance}
\alias{getGraphDistance}
\title{Calculate the graph distance between two sets of images.}
\usage{
getGraphDistance(
  imageList1,
  imageList2,
  isProto1 = FALSE,
  isProto2 = FALSE,
  numPathCuts = 8
)
}
\arguments{
\item{imageList1}{The first image list.}

\item{imageList2}{The second image list.}

\item{isProto1}{Logical. If TRUE, imageList1 is a prototype; otherwise, it's not.}

\item{isProto2}{Logical. If TRUE, imageList2 is a prototype; otherwise, it's not.}

\item{numPathCuts}{The number of path cuts. Default is 8.}
}
\value{
A list containing the graph distance calculation result. The list includes:
\item{matching_weight}{The weight of the optimal edge pairings.}
\item{matching_size}{The size of the optimal edge pairings.}
}
\description{
This function calculates the graph distance between two sets of images represented
by image lists. The graph distance is calculated using linear programming to find
the optimal edge pairings between the two graphs.
}
\details{
This function internally uses \code{\link[handwriter]{getGraphInfo}} to
  find and format the parameter values. It then employs \code{\link[handwriter]{getAllPairsDistances}}
  to calculate the distances between each pair of edges. Finally, the function uses linear programming
  to find the optimal edge pairings and returns the matching weight and size.
}
