% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{getGraphInfo}
\alias{getGraphInfo}
\title{getGraphInfo}
\usage{
getGraphInfo(imageList1, imageList2, isProto1, isProto2, numPathCuts)
}
\arguments{
\item{imageList1}{A graph}

\item{imageList2}{A graph}

\item{isProto1}{True or false. Is the graph information in prototype format?}

\item{isProto2}{True or false. Is the graph information in prototype format?}

\item{numPathCuts}{An integer number of cuts to make when comparing segments of paths}
}
\value{
List of formatted parameters
}
\description{
Gather and format the parameter values need to calculate the distance between
two graphs.
}
\keyword{?}
