% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGraphs.R
\name{getGraphs}
\alias{getGraphs}
\title{Extracts graphs from a .png image using handwriter functions}
\usage{
getGraphs(image, source_folder = getwd(), save_folder = getwd())
}
\arguments{
\item{image}{.png image to be processed}

\item{source_folder}{path to folder containing .png images}

\item{save_folder}{path to folder where graphs are saved to}
}
\value{
saves proclist to an rds file
}
\description{
extractGraphs
}
