% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getNodeGraph}
\alias{getNodeGraph}
\title{getNodeGraph}
\usage{
getNodeGraph(allPaths, nodeList)
}
\arguments{
\item{allPaths}{list of paths}

\item{nodeList}{list of nodes}
}
\value{
a graph of nodes
}
\description{
Internal function for creating a graph from a path list and node list.
}
