% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getNodes}
\alias{getNodes}
\title{getNodes}
\usage{
getNodes(indices, dims)
}
\arguments{
\item{indices}{Where to check for intersection. The indices of pixels locations in 
the thinned image created by thinImage}

\item{dims}{dimensions of the image}
}
\value{
Returns image matrix. 1 is blank, 0 is a node.
}
\description{
Detect intersection points of an image thinned with thinImage.
}
\keyword{Zhang}
\keyword{vertex}
