% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{graphemePaths}
\alias{graphemePaths}
\title{Internal function that uses existing breakPoint list to assign graphemes to the nodes in nodeGraph0.}
\usage{
graphemePaths(allPaths, nodeGraph0, breakPoints)
}
\description{
Internal function that uses existing breakPoint list to assign graphemes to the nodes in nodeGraph0.
}
