% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handwriter-package.R
\docType{package}
\name{handwriter-package}
\alias{handwriter}
\alias{handwriter-package}
\title{handwriter: Handwriting Analysis in R}
\description{
Perform statistical writership analysis of scanned handwritten documents. Webpage provided at: \url{https://github.com/CSAFE-ISU/handwriter}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CSAFE-ISU/handwriter}
  \item Report bugs at \url{https://github.com/CSAFE-ISU/handwriter/issues}
}

}
\author{
\strong{Maintainer}: Stephanie Reinders \email{srein@iastate.edu}

Authors:
\itemize{
  \item Iowa State University of Science and Technology on behalf of its Center for Statistics and Applications in Forensic Evidence [copyright holder, funder]
  \item Nick Berry
  \item James Taylor
}

Other contributors:
\itemize{
  \item Felix Baez-Santiago [contributor]
  \item Jon González [contributor]
}

}
\keyword{internal}
