% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{handwriter_example}
\alias{handwriter_example}
\title{Handwriter Example}
\usage{
handwriter_example(path = NULL)
}
\arguments{
\item{path}{}
}
\value{
Filenames or filepath
}
\description{
List the files in handwriter's "extdata" folder if the path argument is NULL. If
the path argument is the filename of the a file in handwriter's "extdata" folder, 
then the function returns the absolute file path to that file.
}
\examples{
handwriter_example()
handwriter_example("phrase_example.png")
}
