% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_hs_to_i_letter}
\alias{i_hs_to_i_letter}
\title{i_hs_to_i_hs}
\usage{
i_hs_to_i_letter(
  index_nums,
  hs_num_rows,
  letter_num_rows,
  letter_topmost_row,
  letter_leftmost_col
)
}
\arguments{
\item{index_nums}{Vector of index number(s) in handwriting sample}

\item{hs_num_rows}{Integer number of rows in handwriting sample}

\item{letter_num_rows}{Integer number of rows in letter}

\item{letter_topmost_row}{Integer number of the top row of letter}

\item{letter_leftmost_col}{Integer number of the left column of letter}
}
\value{
Integer index number(s) locations in letter
}
\description{
Convert index number(s) locations in a handwriting sample to 
index number(s) locations in a letter in that handwriting sample.
}
