% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_hs_to_rc_hs}
\alias{i_hs_to_rc_hs}
\title{i_hs_to_rc_hs}
\usage{
i_hs_to_rc_hs(index_nums, hs_num_rows)
}
\arguments{
\item{index_nums}{Vector of index number(s) in handwriting sample}

\item{hs_num_rows}{Number of rows in handwriting sample}
}
\value{
List of row and column number(s) locations in handwriting sample
}
\description{
Convert index number(s) locations in a handwriting sample to 
the row and column number(s) locations in that handwriting sample.
}
