% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_to_c}
\alias{i_to_c}
\title{i_to_c}
\usage{
i_to_c(index_num, num_rows)
}
\arguments{
\item{index_num}{Numeric Index number location in matrix}

\item{num_rows}{Number of rows in matrix}
}
\value{
Column number location in matrix
}
\description{
Convert index number location in a matrix to 
the column number location in that matrix.
}
