% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_to_r}
\alias{i_to_r}
\title{i_to_r}
\usage{
i_to_r(index_num, num_rows)
}
\arguments{
\item{index_num}{Numeric Index number location in matrix}

\item{num_rows}{Number of rows in matrix}
}
\value{
Row number location in matrix
}
\description{
Convert index number location in a matrix to 
the row number location in that matrix.
}
\keyword{?}
