% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_to_rc}
\alias{i_to_rc}
\title{i_to_rc}
\usage{
i_to_rc(nodes, dims)
}
\arguments{
\item{nodes}{nodes to be converted.}

\item{dims}{dimensions of binary image}
}
\value{
returns matrix mapping nodes to respective row,
}
\description{
Function for converting indices to respective row, col.
}
\keyword{binary}
\keyword{column}
\keyword{image}
\keyword{row}
