% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{i_to_rci}
\alias{i_to_rci}
\title{i_to_rci}
\usage{
i_to_rci(nodes, dims, fixed = FALSE)
}
\arguments{
\item{nodes}{nodes to be converted.}

\item{img_dim}{dimensions of binary image}
}
\value{
returns matrix mapping nodes' indices to respective row, col
}
\description{
Function for converting indices to respective row, col and associates the original index.
}
\keyword{binary,}
\keyword{column,}
\keyword{image,}
\keyword{index}
\keyword{row,}
