% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{i_to_x}
\alias{i_to_x}
\title{i_to_x}
\usage{
i_to_x(index_num, num_rows)
}
\arguments{
\item{index_num}{Numeric Index number location in matrix}

\item{num_rows}{Number of rows in matrix}
}
\value{
x-coordinate location in matrix
}
\description{
Convert index number location in a matrix to 
the x-coordinate location in that matrix where (0,0)
is the bottom left corner of the matrix
}
