% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFunctions.R
\name{letterToPrototype}
\alias{letterToPrototype}
\title{letterToPrototype}
\usage{
letterToPrototype(letter, numPathCuts = 8)
}
\arguments{
\item{letter}{A graph from a handwriting sample}

\item{numPathCuts}{Number of segments to cut the path(s) into}
}
\value{
List of pathEnds, pathQuarters, and pathCenters given as (x,y) coordinates
with the graph centroid at (0,0). The returned list also contains path lengths.
pathQuarters gives the (x,y) coordinates of the path at the cut points and despite the
name, the path might not be cut into quarters.
}
\description{
Convert the graph to a prototype graph, a graph that serves as a cluster center.
}
\keyword{?}
