% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{line_number_extract}
\alias{line_number_extract}
\title{line_number_extract}
\usage{
line_number_extract(down_dists, all_centroids, dims)
}
\arguments{
\item{down_dists}{how far down to the next character from each character}

\item{all_centroids}{List of centroids extracted from cumulative character_features}

\item{dims}{Dimensions of binary image}
}
\value{
List associating line numbers to characters
}
\description{
Primary logic unit for line number to character association.
}
\keyword{character}
\keyword{features}
\keyword{line}
\keyword{number}
