% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london}
\alias{london}
\title{Cursive written word: London}
\format{
Binary image matrix. 148 rows and 481 columns.
}
\usage{
london
}
\description{
Cursive written word: London
}
\examples{
london_document = list()
london_document$image = london
plotImage(london_document$image)
london_document$thin = thinImage(london_document$image)
plotImageThinned(london_document$image, london_document$thin)
london_processList = processHandwriting(london_document$thin, dim(london_document$image))
}
\keyword{datasets}
