% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{MakeCenterStarts}
\alias{MakeCenterStarts}
\title{MakeCenterStarts}
\usage{
MakeCenterStarts(procList, K, numPathCuts, imagesList)
}
\arguments{
\item{procList}{List of handwriting samples}

\item{K}{Integer number of clusters}

\item{numstrat}{Vector of numbers to sample from each stratum level. The
vector must sum to K.}
}
\value{
centerstarts List of K graphs
}
\description{
Select K graphs as the starting centers of the K clusters. All graphs from
the training handwriting samples are grouped by stratum -- the number of
loops if the graph has one or two loops or the number of paths in the graph.
Then the number of samples specified by numstrat are randomly selected from
each stratum group.
}
\keyword{?}
