% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{makeGifImages}
\alias{makeGifImages}
\title{makeGifImages}
\usage{
makeGifImages(img, img_thin, allPaths, file_path, filenames)
}
\arguments{
\item{allPaths}{List of paths (probably from processHandwriting returned list).}

\item{file_path}{Where the plots created should be saved to. Directory path, relative to current working directory.}

\item{filenames}{Identifying prefix for saved plots. Will be appended with numeric suffixes.}

\item{image}{Binary image.}

\item{image_thin}{Thinned binary image.}
}
\value{
Nothing. Will print out a terminal command for creating your gif. Just navigate to the directory and copy and paste.
}
\description{
Saves a collection of plots, which sequentially display paths in a list, to the provided file path.
Also prints some suggested code for imagemagick gif creation. Can be useful for visualizing path lists or grapheme path lists.
}
\examples{
#### Not Run
# Make csafe_pathList the resulting object from the pathList item in the processHandwriting returned list

makeGifImages(csafe, csafe_thin, csafe_pathList, "../GifPlots/", "csafePaths")

}
