% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{MakeLetterListLetterSpecific}
\alias{MakeLetterListLetterSpecific}
\title{MakeLetterListLetterSpecific}
\usage{
MakeLetterListLetterSpecific(letterList, dims)
}
\arguments{
\item{letterList}{List of letters in a handwriting sample}

\item{dims}{Dimensions of the handwriting sample}
}
\value{
letterList with locations given with respect to each letter
}
\description{
Description
}
\keyword{?}
