% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordSeparation_old.R
\name{makeModel}
\alias{makeModel}
\title{makeModel}
\usage{
makeModel(TaggedJson)
}
\arguments{
\item{TaggedJson}{Json File with tagged letter data}
}
\value{
randomForest model
}
\description{
Creates a randomForest word model
}
