% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{MakeSamplingDF}
\alias{MakeSamplingDF}
\title{MakeSamplingDF}
\usage{
MakeSamplingDF(procList)
}
\arguments{
\item{procList}{List of handwriting samples}
}
\value{
Dataframe
}
\description{
Group each letter in procList by stratum -- the number of loops if the letter has one or two loops
or the number of paths in the letter -- and randomly sample letters from each group. The numstrat
parameter sets the number of letters to sample from each group.
}
