% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeansFunctions.R
\name{meanGraphSet_Kmeans}
\alias{meanGraphSet_Kmeans}
\title{meanGraphSet_Kmeans}
\usage{
meanGraphSet_Kmeans(imList, numPathCuts = 8, meanGraphOrder = "sequential")
}
\arguments{
\item{imList}{List graphs}

\item{numPathCuts}{Integer number of segments to cut each path into}

\item{meanGraphOrder}{Chose the order in which to add graphs to the mean 
graph calculations. The options are 'sequential' and 'random'}
}
\value{
Mean graph
}
\description{
Calculate a mean graph from a set of input graphs
}
\keyword{?}
