% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_wrapped_cauchy}
\alias{model_wrapped_cauchy}
\title{RJAGS Wrapped Cauchy Model}
\format{
Wrapped Cauchy model written in RJAGS syntax.
}
\usage{
model_wrapped_cauchy
}
\description{
RJAGS Wrapped Cauchy Model
}
\examples{
rjags_model <- textConnection(model_wrapped_cauchy)
model_data <- example_model_data
m <- rjags::jags.model(file = rjags_model, data = model_data, n.chains = 1)
}
\keyword{datasets}
