% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nature1}
\alias{nature1}
\title{Full page image of the 4th sample (nature) of handwriting from the first writer.}
\format{
Binary image matrix. 811 rows and 1590 columns.
}
\usage{
nature1
}
\description{
Full page image of the 4th sample (nature) of handwriting from the first writer.
}
\examples{
\dontrun{
nature1_document <- list()
nature1_document$image <- nature1
plotImage(nature1_document)
nature1_document$thin <- thinImage(nature1_document$image)
plotImageThinned(nature1_document)
nature1_processList <- processHandwriting(nature1_document$thin, dim(nature1_document$image))
}
}
\keyword{datasets}
