% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{neighborChanges}
\alias{neighborChanges}
\title{neighborChanges}
\usage{
neighborChanges(coords, img)
}
\arguments{
\item{coords}{Point location.}

\item{img}{Image object.}
}
\value{
Returns count of switches.
}
\description{
Internal function for thinImage. Counts switches from 1 to 0 around a point.
}
