% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{neighbs246468}
\alias{neighbs246468}
\title{neighbs246468}
\usage{
neighbs246468(coords, img)
}
\arguments{
\item{coords}{Point location.}

\item{img}{Image object.}
}
\value{
Returns 1 or 0 for condition mentioned above.
}
\description{
Internal function for thinImage. Returns 1 if one of the points to the top, bottom and left AND one of right, bottom, left are non-zero. 0 otherwise.
}
