% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{neighbs248268}
\alias{neighbs248268}
\title{neighbs246468}
\usage{
neighbs248268(coords, img)
}
\arguments{
\item{coords}{Point location.}

\item{img}{Image object.}
}
\value{
Returns 1 or 0 for condition mentioned above.
}
\description{
Internal function for thinImage. Returns 1 if at least one of the points to the top, right and bottom AND at least one of top, right, left are non-zero. 0 otherwise.
}
