% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{numNeighbors}
\alias{numNeighbors}
\title{numNeighbors}
\usage{
numNeighbors(coords, img)
}
\arguments{
\item{coords}{Point location.}

\item{img}{Image object.}
}
\value{
Returns number of 0 neighbors.
}
\description{
Internal function for thinImage. Counts neighbors with value of 0.
}
