% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{orderNodes}
\alias{orderNodes}
\title{Internal function for ordering the nodes in a letter.}
\usage{
orderNodes(letters, nodesInGraph, dims)
}
\description{
Internal function for ordering the nodes in a letter.
}
