% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{otsuBinarization}
\alias{otsuBinarization}
\title{otsuBinarization}
\usage{
otsuBinarization(img, breaks = 512)
}
\arguments{
\item{img}{image object to be processed}

\item{breaks}{a single number giving the number of cells for the histogram}
}
\value{
separated image into foreground and background
}
\description{
Uses Otsu's Method to binarize given image, performing automatic image thresholding.
}
