% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{pathGraphemeAssociate}
\alias{pathGraphemeAssociate}
\title{Function associating entries in allPaths to each grapheme}
\usage{
pathGraphemeAssociate(allPaths, grapheme)
}
\description{
Function associating entries in allPaths to each grapheme
}
