% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{pathLetterAssociate}
\alias{pathLetterAssociate}
\title{pathLetterAssociate}
\usage{
pathLetterAssociate(allPaths, letter)
}
\arguments{
\item{allPaths}{list of paths}

\item{letter}{individual character}
}
\value{
associated path to each letter
}
\description{
Function associating entries in allPaths to each letter
}
