% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFunctions.R
\name{pathToRC}
\alias{pathToRC}
\title{pathToRC}
\usage{
pathToRC(pathList, dims)
}
\arguments{
\item{pathList}{List of index number locations of paths in a graph}

\item{dims}{Dimensions of the graph's image}
}
\value{
Vector of (x,y) coordinates of the graph's paths
}
\description{
Convert the index number locations of graph's paths to (x,y) coordinates with
(0,0) in the bottom left corner of the graph's image.
}
\keyword{?}
