% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_cluster_fill_counts}
\alias{plot_cluster_fill_counts}
\title{Plot Cluster Fill Counts}
\usage{
plot_cluster_fill_counts(formatted_data, facet = FALSE)
}
\arguments{
\item{formatted_data}{A list of formatted data created by
\code{format_model_data()} or \code{format_questioned_data}.}

\item{facet}{\code{TRUE} uses \code{facet_wrap} to create a subplot for each
writer. \code{FALSE} plots the data on a single plot.}
}
\value{
ggplot plot of cluster fill counts
}
\description{
Plot the cluster fill counts for each document in \code{formatted_data}.
}
\examples{
plot_cluster_fill_counts(formatted_data = example_model_data, facet = TRUE)
plot_cluster_fill_counts(formatted_data = example_questioned_data, facet = FALSE)

}
