% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_graphs}
\alias{plot_graphs}
\title{Plot Graphs}
\usage{
plot_graphs(doc)
}
\arguments{
\item{doc}{A PNG image of handwriting processed with [`processDocument`].}
}
\value{
A plot of all graphs in the document
}
\description{
Plot Graphs
}
\examples{
image_path <- system.file("extdata", "phrase_example.png", package = "handwriter")
doc <- processDocument(image_path)
plot_graphs(doc)

}
