% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace}
\usage{
plot_trace(model, chain, variable)
}
\arguments{
\item{model}{An MCMC object created by \code{\link{fit_model}}}

\item{chain}{An integer chain number}

\item{variable}{The name of a variable in the MCMC object}
}
\value{
ggplot line plot
}
\description{
Create a traceplot for a single variable in an MCMC object created by
\code{\link{fit_model}}.
}
\examples{
\dontrun{
draws <- fit_model(example_model_data, num_iters = 1000, num_chains = 1)
plot_trace(model = draws, chain = 1, variable = "theta[1,1]")
}

}
