% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plotBuckets}
\alias{plotBuckets}
\title{Plot relative bucket frequencies.}
\usage{
plotBuckets(bucketData, plotbucket = "4_BMA")
}
\arguments{
\item{bucketData}{a data matrix with first two columns \code{writer, doc}, then bucket columns to follow. Entries in the data matrix are observed bucket counts.}

\item{plotbucket}{which bucket to plot relative frequencies for.}
}
\description{
Takes a Ybucket (either train or test) from the \code{proclistToData} function and plots relative bucket frequencies broken down by document and writer.
}
