% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterCenters}
\alias{plotClusterCenters}
\title{Plot Template Cluster Centers}
\usage{
plotClusterCenters(template, size = 25)
}
\arguments{
\item{template}{A cluster template created with \code{\link{make_clustering_template}}}

\item{size}{The size of the output plot}
}
\value{
A plot
}
\description{
Plot the cluster centers of a cluster template created with
\code{\link{make_clustering_template}}. The handwriter R package stores the cluster
centers of a cluster template as graph prototypes. A graph prototype consists
of the starting and ending points of each path in the graph, as well as and
evenly spaced points along each path. The prototype also stores the center
point of the graph. All points are represented as xy-coordinates and the
center point is at (0,0).
}
\examples{
# plot cluster centers from example template
plotClusterCenters(example_cluster_template)

}
