% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTemplates_plottingfunctions.R
\name{plotClusterTemplate}
\alias{plotClusterTemplate}
\title{Plot Cluster Template}
\usage{
plotClusterTemplate(template)
}
\arguments{
\item{template}{A cluster template created with \code{\link[=make_clustering_templates]{make_clustering_templates()}}}
}
\value{
A list of plots
}
\description{
Plot the graphs and cluster center in each cluster of a cluster template.
}
\examples{
clusters <- plotClusterTemplate(example_cluster_template)
# show plot of first cluster
clusters[[1]]
# show plot of second cluster
clusters[[2]]

}
