% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotImage}
\alias{plotImage}
\title{plotImage}
\usage{
plotImage(doc)
}
\arguments{
\item{doc}{A document processed with [handwriter::processDocument()] or a binary matrix (all entries are 0 or 1)}
}
\value{
Returns plot of doc$image.
}
\description{
This function plots a basic black and white image.
}
\examples{
csafe_document <- list()
csafe_document$image <- csafe
plotImage(csafe_document)

\dontrun{
document = processDocument('path/to/image.png')
plotImage(document)
}

}
\keyword{plot}
