% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{plotImagePoints}
\alias{plotImagePoints}
\title{ben
probably remove this ://
hmm, takes in list of lists [ [x,y], [x,y], [x,y] .. ] of corrections to be made}
\usage{
plotImagePoints(x)
}
\description{
ben
probably remove this ://
hmm, takes in list of lists [ [x,y], [x,y], [x,y] .. ] of corrections to be made
}
