% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{plotNodesLine}
\alias{plotNodesLine}
\title{plotNodesLine}
\usage{
plotNodesLine(img, thinned, nodeList, nodeSize = 3, nodeColor = "red")
}
\arguments{
\item{img}{full image matrix; used to call plotImageThinned()}

\item{thinned}{thinned image matrix; used to call plotImageThinned()}

\item{nodeList}{list of nodes}

\item{nodeSize}{size of node; default set to 3}

\item{nodeColor}{color of node; default set to red}
}
\value{
a line in between the two nodes
}
\description{
Internal function for drawing a line from two given nodes.
}
