% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{plotPath}
\alias{plotPath}
\title{plotPath}
\usage{
plotPath(path, image, image_thin, zoomBorder = NA, nodeSize = 3)
}
\arguments{
\item{image}{Binary image.}

\item{image_thin}{Thinned binary image.}

\item{zoomBorder}{Pixel padding around the most extreme red triangles in every direction.
Doesn't check boundaries, so be smarter than the plot. Default is NA, which shows the whole image.}

\item{nodeSize}{Size of the red triangles. For fine images a smaller size may be necessary. Default is 3.}
}
\value{
Returns a ggplot2 plot.
}
\description{
Returns a plot with a little red triangle on the pixels in the 'path' vector.
}
\examples{
#### Not Run
# Make csafe_pathList the resulting object from the pathList item in the processHandwriting returned list

plotPath(csafe_pathList[[1]], csafe, csafe_thin, zoomBorder = NA)

}
