% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFunctions.R
\name{pointLineProportionVect}
\alias{pointLineProportionVect}
\title{pointLineProportionVect}
\usage{
pointLineProportionVect(endpt1, endpt2, edgecut_prop_n, edgecutpt_n)
}
\arguments{
\item{endpt1}{Vector of (x,y) coordinates of one endpoint of a path}

\item{endpt2}{Vector of (x,y) coordinates of the other endpoint of a path}

\item{segment_prop_n}{Proportion of the path between the first endpoint and the n-th cut point}

\item{segmentpt_n}{The (x,y) coordinates of the n-th cut point on the path}
}
\value{
Vector of (x,y) coordinates
}
\description{
Subtract the n-th cutpoint on the straight line between the path's endpoints
from the n-th cutpoint on the path itself.
}
\keyword{?}
