% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingUtils.R
\name{preprocess_mutations}
\alias{preprocess_mutations}
\title{preprocess_mutations
expidite the testing process for the new preprocessing algorithm}
\usage{
preprocess_mutations(path, nodesize = 1, init = FALSE)
}
\arguments{
\item{path}{path to binary image, function will handle most of the rest}
}
\description{
preprocess_mutations
expidite the testing process for the new preprocessing algorithm
}
