% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingUtils.R
\name{preprocess_mutations_dif}
\alias{preprocess_mutations_dif}
\title{preprocess_mutations
expidite the testing process for the new preprocessing algorithm}
\usage{
preprocess_mutations_dif(path, clean = TRUE, nodesize = 1,
  init = FALSE)
}
\arguments{
\item{path}{path to binary image, function will handle most of the rest}
}
\description{
preprocess_mutations
expidite the testing process for the new preprocessing algorithm
}
