% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingUtils.R
\name{preprocess_mutations_test}
\alias{preprocess_mutations_test}
\title{preprocess_mutations tester
expidite the testing process for the new preprocessing algorithm
trying reordering some steps of the thinning program}
\usage{
preprocess_mutations_test(path, clean = TRUE, nodesize = 1,
  init = FALSE)
}
\arguments{
\item{path}{path to binary image, function will handle most of the rest}
}
\description{
preprocess_mutations tester
expidite the testing process for the new preprocessing algorithm
trying reordering some steps of the thinning program
}
