% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingUtils.R
\name{preprocessCroppedBinary}
\alias{preprocessCroppedBinary}
\title{preprocessCroppedBinary
This function is intended after crop(), preprocesses images based on masks, uses Python}
\usage{
preprocessCroppedBinary(img, reticulate_init = FALSE, pypath = "./")
}
\arguments{
\item{img}{Cropped binary image (pass through readPNGBinary then crop, found below)}

\item{reticulate_init}{If you do not have reticulate installed, will handle the installation and sourcing of script}

\item{pypath}{path to preprocess.py}
}
\value{
list of values corresponding to fill / clean. 18 sept 18: red = clean, blue = fill, i don't know how to make use of this
}
\description{
preprocessCroppedBinary
This function is intended after crop(), preprocesses images based on masks, uses Python
}
