% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingUtils.R
\name{preprocessToIndexVector}
\alias{preprocessToIndexVector}
\title{preprocessToIndexVector
This function is intended after preprocessCroppedBinary}
\usage{
preprocessToIndexVector(dims, preprocessList)
}
\arguments{
\item{preprocessList}{list resultant after preprocessCroppedBinary}

\item{img}{Cropped binary image (pass through readPNGBinary then crop, found below)}
}
\value{
vector of indicies to plot points via ggplot
}
\description{
preprocessToIndexVector
This function is intended after preprocessCroppedBinary
}
