% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{process_batch_dir}
\alias{process_batch_dir}
\title{Process Batch Directory}
\usage{
process_batch_dir(
  input_dir,
  mask_dir = NULL,
  output_dir = ".",
  skip_docs_on_retry = TRUE
)
}
\arguments{
\item{input_dir}{Input directory that contains images}

\item{mask_dir}{Directory that contains masks for input images}

\item{output_dir}{A directory to save the processed images}

\item{skip_docs_on_retry}{Logical whether to skip documents in input_dir that
caused errors on a previous run. The errors and document names are stored
in output_dir > problems.txt. If this is the first run,
\code{process_batch_list} will attempt to process all documents in input_dir.}
}
\value{
No return value, called for side effects
}
\description{
Process a list of handwriting samples saved as PNG images in a directory:
(1) Load the image and convert it to black and white with \code{\link[=readPNGBinary]{readPNGBinary()}}
(2) Thin the handwriting to one pixel in width with \code{\link[=thinImage]{thinImage()}}
(3) Run \code{\link[=processHandwriting]{processHandwriting()}} to split the handwriting into parts called \emph{edges} and place \emph{nodes} at the ends of
edges. Then combine edges into component shapes called \emph{graphs}.
(4) Save the processed document in an RDS file.
(5) Optional. Return a list of the processed documents.
}
\examples{
\dontrun{
process_batch_dir("path/to/input_dir", "path/to/output_dir")
}

}
