% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_writing.R
\name{processHandwriting}
\alias{processHandwriting}
\title{Process Handwriting by Component}
\usage{
processHandwriting(img, dims)
}
\arguments{
\item{img}{Thinned binary image created with \code{\link[=thinImage]{thinImage()}}.}

\item{dims}{Dimensions of thinned binary image.}
}
\value{
A list of the processed image
}
\description{
The main driver of handwriting processing. Takes in an image of thinned
handwriting created with \code{\link[=thinImage]{thinImage()}} and splits the the handwriting into
shapes called \emph{graphs}. Instead of processing the entire document at once,
the thinned writing is separated into connected components and each component
is split into graphs.
}
\examples{
twoSent_document <- list()
twoSent_document$image <- twoSent
twoSent_document$thin <- thinImage(twoSent_document$image)
twoSent_processList <- processHandwriting(twoSent_document$thin, dim(twoSent_document$image))

}
