% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{qdocs_normslope}
\alias{qdocs_normslope}
\title{Prepare test data for posterior analysis}
\usage{
qdocs_normslope(bucketData, measData, writers = 1:27, writerIndices = TRUE)
}
\arguments{
\item{bucketData}{test bucket data from same \code{getdata_mb} call as train data}

\item{writers}{a vector of writers, if \code{writerIndices = TRUE} then this is a vector of indices, otherwise it is writer IDs.}

\item{writerIndices}{see \code{writers} above}
}
\description{
Prepare test data for posterior analysis
}
