% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{rc_hs_to_rc_letter}
\alias{rc_hs_to_rc_letter}
\title{rc_hs_to_rc_letter}
\usage{
rc_hs_to_rc_letter(row_nums, col_nums, letter_topmost_row, letter_leftmost_col)
}
\arguments{
\item{row_nums}{Integer row numbers in handwriting sample}

\item{col_nums}{Integer column numbers in handwriting sample}

\item{letter_topmost_row}{Integer number of the top row of letter}

\item{letter_leftmost_col}{Integer number of the left column of letter}
}
\value{
List of row and column number(s) locations in letter
}
\description{
Convert row and column number(s) locations in a handwriting sample to 
row and column number(s) locations in a letter in that handwriting sample.
}
