% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{rc_to_i}
\alias{rc_to_i}
\title{rc_to_i}
\usage{
rc_to_i(row_y, col_x, img_dim, fixed = FALSE)
}
\arguments{
\item{row_y}{Row(s) to be converted to an index}

\item{col_x}{Columns(s) to be converted to an index}

\item{img_dim}{Dimensions of binary image}
}
\value{
Returns index(icies) of all row_y's and col_x's
}
\description{
Convert rows and columns to their respective indices.
This is index sensitive, so row_y[[1]] should correspond to col_x[[1]]
}
\keyword{binary,}
\keyword{column,}
\keyword{image,}
\keyword{index}
\keyword{row,}
