% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgb2grayscale}
\alias{rgb2grayscale}
\title{rgba2rgb}
\usage{
rgb2grayscale(img)
}
\arguments{
\item{img}{A 3-d array with slices R, G, B, and alpha.}
}
\description{
Removes alpha channel from png image.
}
