% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgba2rgb}
\alias{rgba2rgb}
\title{rgba2rgb}
\usage{
rgba2rgb(img)
}
\arguments{
\item{img}{A 3-d array with slices R, G, B, and alpha.}
}
\value{
img as a 3D array with alpha channel removed
}
\description{
Removes alpha channel from png image.
}
