% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHandwritingViewer.R
\name{runHandwritingViewer}
\alias{runHandwritingViewer}
\title{runHandwritingViewer}
\usage{
runHandwritingViewer()
}
\value{
None
}
\description{
This function opens and runs a shiny app that allows for viewing of an object 
that comes from the `processHandwriting` function.` Requires \pkg{\link{shiny}}.
}
\seealso{
\pkg{\link{lattice}}
}
