% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeansFunctions.R
\name{runLetterKmeans}
\alias{runLetterKmeans}
\title{runLetterKmeans}
\usage{
runLetterKmeans(
  imList,
  centers,
  K,
  numberToRun,
  numCores = 1,
  numDistCores = 0,
  numPathCuts = 8,
  iter.max = 500,
  numOutliers = round(0.25 * length(imList)),
  gamma = 3,
  results_file,
  meanGraph = "slow_change",
  meanGraphOrder = "sequential",
  fillClusters = "farthest"
)
}
\arguments{
\item{imList}{List graphs from handwriting samples}

\item{centers}{List of starting cluster centers}

\item{K}{Integer number of clusters. If numOutliers > 0, then
there will be K+1 clusters.}

\item{numberToRun}{Integer number of clustering templates to create}

\item{numCores}{Integer number of cores. Each clustering template will be created on
a different core}

\item{numDistCores}{Integer number of cores to use for distance calculations}

\item{numPathCuts}{Integer number of segments to cut each path into}

\item{iter.max}{Integer maximum number of iterations to allow cluster centers to converge}

\item{numOutliers}{Integer number of outliers to allow. If numOutliers is zero, no outliers will be
assigned.}

\item{gamma}{Float parameter for calculating the outlier cutoff. If numOutliers is zero, gamma has no effect.}

\item{results_file}{String name of file without the file extension to save the clustering template}

\item{meanGraph}{String choice of algorithm for calculating mean graphs. The choices are 'basic',
'slow_change' and 'kmeans'}

\item{meanGraphOrder}{Chose the order in which to add graphs to the mean 
graph calculations. The options are 'sequential' and 'random'}

\item{fillClusters}{String choice of how to choose which graph to reassign to an empty cluster. 
Choice are 'random' and 'farthest'}
}
\value{
List of clustering template(s)
}
\description{
Create and save one or more clustering templates.
}
\keyword{?}
