% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCSAFEDocs.R
\name{select_csafe_docs}
\alias{select_csafe_docs}
\title{select_csafe_docs}
\usage{
select_csafe_docs(
  num_template_writers,
  template_sessions,
  template_reps,
  template_prompts,
  template_seed,
  num_model_writers,
  model_sessions,
  model_reps,
  model_prompts,
  model_seed,
  questioned_sessions,
  questioned_reps,
  questioned_prompts
)
}
\arguments{
\item{num_template_writers}{An integer number of writers to randomly select
for cluster template creation.}

\item{template_sessions}{Vector of writing session(s) to use for cluster template creation}

\item{template_reps}{Vector of writing repetition(s) to use for cluster template creation}

\item{template_prompts}{Vector of writing prompt(s) to use for cluster template creation}

\item{template_seed}{An integer seed for randomly selecting writers to use for cluster template creation}

\item{num_model_writers}{An integer number of writers to randomly select
for fitting a model}

\item{model_sessions}{Vector of writing session(s) to use for fitting a model}

\item{model_reps}{Vector of writing repetitions(s) to use for fitting a model}

\item{model_prompts}{Vector of writing prompt(s) to use for fitting a model}

\item{model_seed}{An integer seed for randomly selecting writers to use for fitting a model}

\item{questioned_sessions}{Vector of writing session(s) to use for questioned (test) documents}

\item{questioned_reps}{Vector of writing repetitions(s) to use for questioned (test) documents}

\item{questioned_prompts}{Vector of writing prompts(s) to use for questioned (test) documents}
}
\value{
A list of data frames
}
\description{
Build three sets of documents from the CSAFE Handwriting Database: cluster
template training documents; model training documents; and questioned (test)
documents. The CSAFE Handwriting Database contains 200+ writers. Each writer
wrote three prompts: the London letter, a common phrase, and an excerpt from
the Wizard of Oz. Each writer wrote each prompt three times during three
different sessions. Each session was at least three weeks apart.
}
\details{
The \code{select_csafe_docs} function allows the user to select the number of
writers to use for the template and model, and ensures that no writers are
used in both sets. The model implemented in handwriter is only applicable to
closed-sets of writers: the questioned document is known to have been written
by one writers whose known handwriting samples are used to fit the model. The
\code{select_csafe_docs} function selects handwriting samples from the model
writers for use as questioned documents. While the writers are the same, the
a document cannot be used in both the model and the questioned sets. If a
user attempts to do this, \code{select_csafe_docs} will return an error message.

The \code{select_csafe_docs} function also allows the user to choose which
prompt(s), session(s), and repetition(s) to use for each of the three sets.
As already mentioned, the function won't allow the user to select the same
documents for the model and questioned sets.
}
\examples{
docs <- select_csafe_docs(
  num_template_writers = 10,
  template_sessions = 1,
  template_reps = 1,
  template_prompts = "London Letter",
  template_seed = 100,
  num_model_writers = 5,
  model_sessions = 1,
  model_reps = c(1, 2, 3),
  model_prompts = "Wizard of Oz",
  model_seed = 101,
  questioned_sessions = 3,
  questioned_reps = 1,
  questioned_prompts = "Wizard of Oz"
)

}
