% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCSAFEDocs.R
\name{select_model_docs}
\alias{select_model_docs}
\title{select_model_docs}
\usage{
select_model_docs(template_df, num_writers, sessions, reps, prompts, seed)
}
\arguments{
\item{template_df}{A data frame of template training documents created with
[`select_template_docs`]}

\item{num_writers}{Integer number of writers to randomly select}

\item{sessions}{Vector of writing session(s) to use}

\item{reps}{Vector of writing repetition(s) to use}

\item{prompts}{Vector of writing prompt(s) to use}

\item{seed}{Integer seed for the random number generator}
}
\value{
A data frame of selected documents
}
\description{
Randomly select model training documents to fit a model. The model writers
will be distinct from the template training writers.
}
