% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCSAFEDocs.R
\name{select_questioned_docs}
\alias{select_questioned_docs}
\title{select_questioned_docs}
\usage{
select_questioned_docs(model_df, sessions, reps, prompts)
}
\arguments{
\item{model_df}{A data frame of model training documents created with
[`select_model_docs`]}

\item{sessions}{Vector of writing session(s) to use}

\item{reps}{Vector of writing repetition(s) to use}

\item{prompts}{Vector of writing prompt(s) to use}
}
\value{
A data frame of selected documents
}
\description{
Randomly select questioned documents from the same writers used to fit the model.
}
