% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCSAFEDocs.R
\name{select_template_docs}
\alias{select_template_docs}
\title{select_template_docs}
\usage{
select_template_docs(num_writers, sessions, reps, prompts, seed)
}
\arguments{
\item{num_writers}{Integer number of writers to randomly select}

\item{sessions}{Vector of writing session(s) to use}

\item{reps}{Vector of writing repetition(s) to use}

\item{prompts}{Vector of writing prompt(s) to use}

\item{seed}{Integer seed for the random number generator}
}
\value{
A data frame of selected documents
}
\description{
Randomly select template training documents to create a clustering template.
}
