% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{solveLP}
\alias{solveLP}
\title{solveLP}
\usage{
solveLP(dists)
}
\arguments{
\item{endpt1}{Vector of (x,y) coordinates of one endpoint of a path}

\item{endpt2}{Vector of (x,y) coordinates of the other endpoint of a path}

\item{segment_prop_n}{Proportion of the path between the first endpoint and
the n-th cut point}

\item{segmentpt_n}{The (x,y) coordinates of the n-th cut point on the path}
}
\value{
Vector of (x,y) coordinates
}
\description{
Use linear programming to solve a constrained minimization optimization
problem to find the optimal pairings of paths between two graphs.
}
\keyword{?}
