% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{stepA}
\alias{stepA}
\title{stepA}
\usage{
stepA(img, coords)
}
\arguments{
\item{img}{Image object.}

\item{coords}{Locations of points to check.}
}
\value{
TRUE or FALSE for each point in coords, based on if they pass the checks.
}
\description{
Internal function for thinImage. First iteration of checks for thinning algorithm
}
