% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{stepB}
\alias{stepB}
\title{stepA}
\usage{
stepB(img, coords)
}
\arguments{
\item{img}{Image object.}

\item{coords}{Locations of points to check.}
}
\value{
TRUE or FALSE for each point in coords, based on if they pass the checks.
}
\description{
Internal function for thinImage. Second iteration of checks for thinning algorithm
}
