% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{thinImage}
\alias{thinImage}
\title{thinImage}
\usage{
thinImage(img, verbose = FALSE)
}
\arguments{
\item{img}{Binary Image matrix.}

\item{verbose}{Whether you want detailed output as code is run. Defaults to FALSE.}
}
\value{
Returns thinned image matrix.
}
\description{
Fast implementation of the Zhang-Suen Thinning algorithm (1984).
}
\examples{
data(london)
london = crop(london)
london_thin = thinImage(london, verbose = TRUE)


data(message)
message = crop(message)
message_thin = thinImage(message, verbose = TRUE)

}
\keyword{Suen}
\keyword{Zhang,}
\keyword{thinning,}
