% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoSent}
\alias{twoSent}
\title{Two sentence printed example handwriting}
\format{
Binary image matrix. 396 rows and 1947 columns
}
\usage{
twoSent
}
\description{
Two sentence printed example handwriting
}
\examples{
twoSent_document <- list()
twoSent_document$image <- twoSent
plotImage(twoSent_document)

\dontrun{
twoSent_document <- list()
twoSent_document$image <- twoSent
plotImage(twoSent_document)
twoSent_document$thin <- thinImage(twoSent_document$image)
plotImageThinned(twoSent_document)
twoSent_processList <- processHandwriting(twoSent_document$thin, dim(twoSent_document$image))
}
}
\keyword{datasets}
