% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{weightedMeanGraphs}
\alias{weightedMeanGraphs}
\title{weightedMeanGraphs}
\usage{
weightedMeanGraphs(
  imageList1,
  imageList2,
  p1,
  isProto1 = FALSE,
  isProto2 = FALSE,
  numPathCuts = 8
)
}
\arguments{
\item{imageList1}{List of images}

\item{imageList2}{List of images}

\item{p1}{?}

\item{isProto1}{Whether imageList1 is exemplars}

\item{isProto2}{Whether imageList2 is exemplars}

\item{numPathCuts}{Number of cuts}
}
\value{
?
}
\description{
Description
}
