% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{whichNeighbors}
\alias{whichNeighbors}
\title{Internal function for identifying which neighbors are black.}
\usage{
whichNeighbors(coords, img)
}
\description{
Internal function for identifying which neighbors are black.
}
