% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whichToFill}
\alias{whichToFill}
\title{whichToFill}
\usage{
whichToFill(img)
}
\arguments{
\item{img}{A binary matrix.}
}
\value{
A cleaned up image.
}
\description{
Finds pixels in the plot that shouldn't be white and makes them black. 
Quick and helpful cleaning for before the thinning algorithm runs.
}
