% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\docType{data}
\name{wordModel}
\alias{wordModel}
\title{wordModel is the RandomForest model to decide if a word separation has happened}
\description{
wordModel is the RandomForest model to decide if a word separation has happened
}
\keyword{data}
