% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordSeparation_old.R
\docType{data}
\name{wordModelNew}
\alias{wordModelNew}
\title{wordModelNew is the RandomForest model to decide if a word separation has happened}
\description{
wordModelNew is the RandomForest model to decide if a word separation has happened
}
\keyword{data}
