% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_model.R
\name{about_variable}
\alias{about_variable}
\title{About Varialbe}
\usage{
about_variable(variable, model)
}
\arguments{
\item{variable}{A variable in the fitted model output by \code{\link[=fit_model]{fit_model()}}}

\item{model}{A fitted model created by \code{\link[=fit_model]{fit_model()}}}
}
\value{
Text that explains the variable
}
\description{
\code{about_variable()} returns information about the model variable.
}
\examples{
about_variable(
  variable = "mu[1,2]",
  model = example_model
)

}
