% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{add_word_info}
\alias{add_word_info}
\title{add_word_info}
\usage{
add_word_info(letterList, dims)

add_word_info(letterList, dims)
}
\arguments{
\item{character_features}{All extracted features}

\item{img_dim}{Dimensions of binary image}
}
\value{
Appends line information to character features

Appends line information to character features
}
\description{
Associates characters to their respective word numbers by distance between right edge of char and left edge of next
Needs improvement if runtime becomes a problem

Associates characters to their respective word numbers by ML on labeled data
}
\keyword{character,}
\keyword{features,}
\keyword{line}
\keyword{number}
