% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordSeparation_old.R
\name{add_word_info}
\alias{add_word_info}
\title{add_word_info}
\usage{
add_word_info(letterList, dims)
}
\arguments{
\item{letterList}{List containing characters}

\item{dims}{Dimensions of binary image}
}
\value{
Appends line information to character features
}
\description{
Associates characters to their respective word numbers by ML on labeled data
}
\keyword{character}
\keyword{features}
\keyword{line}
\keyword{number}
