% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{AddLetterImages}
\alias{AddLetterImages}
\title{AddLetterImages}
\usage{
AddLetterImages(letterList, docDims)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{docDims}{Dimensions of the original document}
}
\value{
letterList with a new matrix `image` value for each sublist.
}
\description{
This function returns a plot of a single letter extracted from a document. It uses the letterList parameter from the processHandwriting function and accepts a single value as whichLetter. Dims requires the dimensions of the entire document, since this isn't contained in processHandwriting.
}
