% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{AllUniquePaths}
\alias{AllUniquePaths}
\title{AllUniquePaths}
\usage{
AllUniquePaths(adj, graph, graph0)
}
\arguments{
\item{adj}{adjacent matrix}

\item{graph}{first skeletonized graph}

\item{graph0}{second skeletonized graph}
}
\value{
a list of all non loop paths
}
\description{
Internal function for getting a list of all non loop paths in a writing sample.
}
