% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{analyze_questioned_documents}
\alias{analyze_questioned_documents}
\title{analyze_questioned_documents}
\usage{
analyze_questioned_documents(model_data, model, questioned_data, num_cores)
}
\arguments{
\item{model_data}{A list of input data formatted with
\code{format_model_data} for rjags}

\item{model}{A list of MCMC draws from a model fit with \code{\link[=fit_model]{fit_model()}}.}

\item{questioned_data}{A list of questioned documents' data formatted with
\code{format_questioned_data()}.}

\item{num_cores}{An integer number of cores to use for parallel processing
with the \code{doParallel} package.}
}
\value{
A list of likelihoods, votes, and posterior probabilities of
writership for each questioned document.
}
\description{
\code{analyze_questioned_documents()} estimates the posterior probability of
writership for the questioned documents using MCMC draws from a hierarchical
model created with \code{fit_model()}.
}
\examples{
model <- fit_model(
  model_data = example_model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(model = model, burn_in = 250)

}
\keyword{model}
