% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{char_to_feature}
\alias{char_to_feature}
\title{char_to_feature}
\usage{
char_to_feature(character, img_dim, uniqueid)
}
\arguments{
\item{character}{character to extract information from}

\item{img_dim}{Dimensions of binary image}

\item{uniqueid}{Unique numerical reference to character}
}
\value{
List containing features of character
}
\description{
Secondary driver of feature extraction
Extracts features from a single character
}
\keyword{character,}
\keyword{features}
