% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{checkBreakPoints}
\alias{checkBreakPoints}
\title{Internal function called by processHandwriting that eliminates breakpoints based on rules to try to coherently separate letters}
\usage{
checkBreakPoints(candidateNodes, allPaths, nodeGraph, terminalNodes, dims)
}
\description{
Internal function called by processHandwriting that eliminates breakpoints based on rules to try to coherently separate letters
}
