% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{checkStacking}
\alias{checkStacking}
\title{CheckStacking}
\usage{
checkStacking(candidateBreaks, allPaths, letters, nodeGraph0, dims)
}
\arguments{
\item{candidateBreaks}{possible breaks for letterpath}

\item{allPaths}{list of paths}

\item{letters}{list of letters}

\item{nodeGraph0}{skeletonized graph}

\item{dims}{graph dimensions}
}
\description{
Internal function for removing breakpoints that follow all of the rules, but separate two letters that are
stacked on top of eachother. Currently, this is done in a very ad hoc, and untested manner. Will look for a better
solution in the future.
}
