% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cleanBinaryImage}
\alias{cleanBinaryImage}
\title{cleanBinaryImage}
\usage{
cleanBinaryImage(img)
}
\arguments{
\item{img}{A matrix of 1s and 0s.}
}
\value{
png image with the alpha channel removed
}
\description{
Removes alpha channel from png image.
}
