% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{countChanges}
\alias{countChanges}
\title{countChanges}
\usage{
countChanges(coords, img)
}
\arguments{
\item{coords}{\(row, column\) coordinates of a single point to consider}

\item{img}{The non-thinned image as binary bit map. Double-check: processHandwriting might call countChanges on thinned image}
}
\value{
The number of edges connected to coords
}
\description{
Internal function for counting the number of edges connected to a point
}
