% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{countChanges}
\alias{countChanges}
\title{countChanges}
\usage{
countChanges(coords, img)
}
\arguments{
\item{coords}{coordinates to consider}

\item{img}{The non-thinned image as binary bit map}
}
\value{
The sum of the 4-connected components around a pixel.
}
\description{
Internal function for counting 4-connected components around a pixel.
}
