% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateWords.R
\name{create_words}
\alias{create_words}
\title{create_words}
\usage{
create_words(processList)
}
\arguments{
\item{processList}{Output from processHandwriting - contains all glyph information}
}
\value{
list of word objects
}
\description{
creates word objects based on splits found in processHandwriting
}
\examples{
twoSent_document = list()
twoSent_document$image = twoSent
twoSent_document$thin = thinImage(twoSent_document$image)
twoSent_processList = processHandwriting(twoSent_document$thin, dim(twoSent_document$image))

dims = dim(twoSent_document$image)
words = create_words(twoSent_processList) 
words_after_processing = process_words(words, dim(twoSent_document$image), TRUE)
}
