% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csafe}
\alias{csafe}
\title{Cursive written word: csafe}
\format{
Binary image matrix. 111 rows and 410 columns.
}
\usage{
csafe
}
\description{
Cursive written word: csafe
}
\keyword{datasets}
