% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{drop_burnin}
\alias{drop_burnin}
\title{drop_burnin}
\usage{
drop_burnin(draws, burn_in)
}
\arguments{
\item{draws}{A list of dataframes of MCMC draws created by \code{fit_model()}}

\item{burn_in}{An integer number of starting iterations of MCMC to drop.}
}
\value{
A list of data frames of MCMC draws with burn-in dropped.
}
\description{
\code{drop_burnin()} removes the burn-in from the MCMC draws.
}
\examples{
draws <- fit_model(example_model_training_data, 4000)
draws <- drop_burnin(draws, 1000)

}
\keyword{model}
