% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model_clusters}
\alias{example_model_clusters}
\title{Example of list of model training data for the hierarchical model}
\format{
List created by \code{\link{get_clusterassignment}} with extra fields removed to
make the example file smaller.
}
\usage{
example_model_clusters
}
\description{
Example of list of model training data for the hierarchical model
}
\examples{
writer_indices <- c(2, 5)
doc_indices <- c(7, 18)
model_data <- format_model_data(example_model_clusters,
  writer_indices,
  doc_indices,
  a = 2, b = 0.25, c = 2, d = 2, e = 0.5
)
}
\keyword{datasets}
