% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_clusters}
\alias{example_questioned_clusters}
\title{Example questioned documents with with cluster assignments.}
\format{
List created by \code{\link{get_clusterassignment}} with extra fields removed
to make the example file smaller.
}
\usage{
example_questioned_clusters
}
\description{
The example questioned documents are included in the package. They are
located in \code{system.file("extdata/example_images/template_training_images", package = "handwriter")}. The documents were first processed with
\code{\link{process_batch_dir}} then their processed graphs where assigned to the
nearest cluster in the example cluster template with
\code{\link{get_clusterassignment}}.
}
\examples{
# format questioned data
questioned_data <- format_questioned_data(
  formatted_model_data = example_model_data,
  questioned_proc_list = example_questioned_clusters,
  writer_indices = c(2, 5),
  doc_indices = c(7, 18)
)

# plot cluster fill counts
plot_cluster_fill_counts(questioned_data)

# analyze questioned documents
analysis <- analyze_questioned_documents(
  model_data = example_model_data,
  model = example_model_1chain,
  questioned_data = example_questioned_data,
  num_cores = 2
)
analysis$posterior_probabilities

}
\keyword{datasets}
