% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_clusters}
\alias{example_questioned_clusters}
\title{Example questioned documents with with cluster assignments.}
\format{
List created by \code{\link{get_clusterassignment}} with extra fields removed
to make the example file smaller.

List created by \code{\link{get_clusterassignment}} with extra fields removed to
make the example file smaller.
}
\usage{
example_questioned_clusters

example_questioned_clusters
}
\description{
The example questioned documents are included in the package. They are
located in \code{system.file("extdata/example_images/template_training_images", package = "handwriter")}. The documents were first processed with
\code{\link{process_batch_dir}} then their processed graphs where assigned to the
nearest cluster in the example cluster template with
\code{\link{get_clusterassignment}}.
}
\examples{
writer_indices <- c(2, 5)
doc_indices <- c(7, 18)
model_data <- format_questioned_data(
  example_questioned_clusters,
  writer_indices,
  doc_indices
)
}
\keyword{datasets}
