% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_questioned_data}
\alias{example_questioned_data}
\title{Example of questioned data formatted for the hierarchical model}
\format{
A named list created by \code{\link{format_questioned_data}} with 2 items:
\describe{
\item{graph_measurements}{data frame that shows the writer ID,
document ID, slope, principal component rotation angle, and wrapped
principal rotation angle for each graph }
\item{cluster_fill_counts}{data frame that shows the number of graphs
assigned to each cluster for each document.}
}
}
\usage{
example_questioned_data
}
\description{
Example of questioned data formatted for the hierarchical model
}
\examples{
draws <- fit_model(example_model_data, 4000)
analysis <- analyze_questioned_documents(example_model_data,
  draws,
  example_questioned_data,
  num_cores = 4
)

}
\keyword{datasets}
