% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{extract_character_features}
\alias{extract_character_features}
\title{extract_character_features}
\usage{
extract_character_features(img, character_lists, dims)
}
\arguments{
\item{img}{The thinned image bitmap}

\item{character_lists}{Output from processHandwriting$letterLists}

\item{dims}{Dimensions of binary image}
}
\value{
nested lists associating features to respective characters.
}
\description{
Primary driver of feature extraction. Parses all characters from a processed image.
}
\keyword{centroid,}
\keyword{character}
\keyword{lean,}
\keyword{skew,}
\keyword{slant,}
