% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{findMergeNodes}
\alias{findMergeNodes}
\title{Internal function to merge nodes that are very close together.}
\usage{
findMergeNodes(skel_graph, mergeMat)
}
\description{
Internal function to merge nodes that are very close together.
}
