% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(model_data, num_iters, num_chains = 1)
}
\arguments{
\item{model_data}{A list of input data formatted with
\code{format_model_data} for rjags.}

\item{num_iters}{An integer number of iterations of MCMC.}

\item{num_chains}{An integer number of chains to use.}
}
\value{
A list of data frames of MCMC draws.
}
\description{
\code{fit_model()} fits a Bayesian hierarchical model to \code{model_training_data} and
draws samples from the model with MCMC.
}
\examples{
model <- fit_model(
  model_data = example_model_data,
  num_iters = 500,
  num_chains = 1
)
model <- drop_burnin(model = model, burn_in = 250)

}
\keyword{model}
