% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_modelfunctions.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(model_training_data, num_iters)
}
\arguments{
\item{model_training_data}{A list of input data formatted with
\code{format_model_data} for rjags}

\item{num_iters}{An integer number of iterations of MCMC.}
}
\value{
A list of data frames of MCMC draws.
}
\description{
\code{fit_model()} fits a Bayesian hierarchical model to \code{model_training_data} and
draws samples from the model with MCMC.
}
\examples{
draws <- fit_model(example_model_data$rjags_data, 1000)

}
\keyword{model}
