% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{format_model_data}
\alias{format_model_data}
\title{format_model_data}
\usage{
format_model_data(proc_list, writer_indices, doc_indices, a, b, c, d, e)
}
\arguments{
\item{proc_list}{List of processed handwriting from a set of documents created by
\code{get_clusterassignment()}. Each item in the list contains the extracted graphs and from a document.}

\item{writer_indices}{Vector of start and end indices for the writer id in the document names.}

\item{doc_indices}{Vector of start and end indices for the document id in the document names.}

\item{a}{Scalar}

\item{b}{Scalar}

\item{c}{Scalar}

\item{d}{Scalar}

\item{e}{Scalar}
}
\value{
List of data formated for rjags.
}
\description{
\code{format_model_data()} formats the data need for the rjags model.
}
