% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{format_questioned_data}
\alias{format_questioned_data}
\title{format_questioned_data}
\usage{
format_questioned_data(proc_list, writer_indices, doc_indices)
}
\arguments{
\item{proc_list}{List of processed handwriting from a set of questioned documents
created by \code{get_clusterassignment()}. Each item in the list contains the
extracted graphs from a document.}

\item{writer_indices}{Vector of start and end indices for the writer id in
the document names.}

\item{doc_indices}{Vector of start and end indices for the document id in the
document names.}
}
\value{
List of data formatted analysis.
}
\description{
\code{format_questioned_data()} formats the questioned data for analysis with the hierarchical model.
}
\examples{
writer_indices = c(2,5)
doc_indices = c(7,18)
format_questioned_data(example_model_proc_list, writer_indices, doc_indices)

}
