% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_aspect_info}
\alias{get_aspect_info}
\title{get_aspect_info}
\usage{
get_aspect_info(character, dims)
}
\arguments{
\item{character}{character to extract information from}

\item{dims}{Dimensions of binary image}
}
\value{
List containing aspect_ratio, 
height, width, highest, lowest, leftmost,
points of character. (row, col representation)
}
\description{
Extracts aspect ratio & supporting information from a character
Relevant Features:
Aspect Ratio: Row (Height) over (Column Width) //Determined after a meeting, can easily be switched around)
Height, Width (Each measure of pixels)
The rest are supporting features that are minor independently.
}
\keyword{aspect,}
\keyword{character,}
\keyword{height}
\keyword{ratio,}
\keyword{width,}
