% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_centroid_info}
\alias{get_centroid_info}
\title{get_centroid_info}
\usage{
get_centroid_info(character, img_dim)
}
\arguments{
\item{character}{character to extract information from}

\item{img_dim}{Dimensions of binary image}
}
\value{
List containing centroid, pixel density,
letter 'lean', and all supporting information
}
\description{
Extracts centroid & supporting information from a character
}
\keyword{centroid,}
\keyword{character}
\keyword{lean,}
\keyword{skew,}
\keyword{slant,}
