% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_datafunctions.R
\name{get_cluster_fill_counts}
\alias{get_cluster_fill_counts}
\title{get_cluster_fill_counts}
\usage{
get_cluster_fill_counts(proc_list, writer_indices, doc_indices)
}
\arguments{
\item{proc_list}{List of processed handwriting from a set of documents
created by `get_clusterassignment()`. Each item in the list contains the
extracted graphs from a document.}

\item{writer_indices}{Vector of start and end indices for the writer id in
the document names.}

\item{doc_indices}{Vector of start and end indices for the document id in the
document names.}
}
\value{
A data frame of cluster fill counts for each document in `proc_list`.
}
\description{
`get_cluster_fill_counts` counts the number of graphs assigned to each
cluster in the cluster template for each document in `proc_list`.
}
