% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_loop_info}
\alias{get_loop_info}
\title{get_loop_info}
\usage{
get_loop_info(character, dims)
}
\arguments{
\item{character}{Target for loop association}

\item{dims}{Dimensions of binary image}
}
\value{
Loop information to respective character
}
\description{
Associator of loop to character association
Relevant Features:
Loop Count, how many loops are found in the letter
Loop Major, length of farthest line that can be drawn inside of a loop
Loop Minor, length of the perpendicular bisector of the loop major.
}
