% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{get_loop_info}
\alias{get_loop_info}
\title{}
\usage{
get_loop_info(character, img_dim)
}
\arguments{
\item{character}{Target for loop association}

\item{img_dim}{Dimensions of binary image}
}
\value{
Loop information to respective character
}
\description{
Associator of loop to character association
Relevant Features:
Loop Count, how many loops are found in the letter
Loop Major, length of farthest line that can be drawn inside of a loop
Loop Minor, length of the perpindcular bisector of the loop major.
! I've removed loop minor / loop major features due to instability during testing.
The outliers were frequent enough to deem my implementation unreliable for modeling.
I need additional help in some of the syntactical challenges I've experienced with those features..
}
\keyword{associate}
\keyword{character,}
\keyword{loop,}
