% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFunctions.R
\name{getGraphInfo}
\alias{getGraphInfo}
\title{getGraphInfo}
\usage{
getGraphInfo(imageList1, imageList2, isProto1, isProto2, numPathCuts)
}
\arguments{
\item{imageList1}{A graph}

\item{imageList2}{A graph}
}
\value{
List of formatted parameters
}
\description{
Gather and format the parameter values need to calculate the distance between
two graphs.
}
\keyword{?}
