% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getLoops}
\alias{getLoops}
\title{Internal function for getting looped paths.}
\usage{
getLoops(nodeList, graph, graph0, pathList, dims)
}
\description{
Internal function for getting looped paths.
}
