% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getLoops}
\alias{getLoops}
\title{Internal function for getting looped paths.}
\usage{
getLoops(nodeList, graph, graph0, pathList, dims)
}
\arguments{
\item{nodeList}{A list of all found nodes}

\item{graph}{first skeletonized graph}

\item{graph0}{second skeletonized graph}

\item{pathList}{The current path list to check for loops}

\item{dims}{dimensions of the image}
}
\description{
Internal function for getting looped paths.
}
