% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getNodeGraph}
\alias{getNodeGraph}
\title{Internal function for creating a graph from a path list and node list.}
\usage{
getNodeGraph(allPaths, nodeList)
}
\description{
Internal function for creating a graph from a path list and node list.
}
