% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getNodeOrder}
\alias{getNodeOrder}
\title{Internal function for ordering the nodes in a letter.}
\usage{
getNodeOrder(letter, nodesInGraph, nodeConnectivity, dims)
}
\description{
Internal function for ordering the nodes in a letter.
}
