% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{getNodes}
\alias{getNodes}
\title{getNodes}
\usage{
getNodes(indices, dims)
}
\arguments{
\item{img}{Thinned binary image.}
}
\value{
Returns image matrix. 1 is blank, 0 is a node.
}
\description{
Detect intersection points of an image thinned with thinImage.
}
\examples{
data(london)
london = crop(london)
london_thin = thinImage(london)
london_nodes = getNodes(london_thin, dim(london))

## Not Run
#data(message)
#message = crop(message)
#message_thin = thinImage(message)
#message_nodes = getNodes(message_thin, dim(message))

}
\keyword{Suen}
\keyword{Zhang,}
\keyword{detection,}
\keyword{vertex}
