% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handwriter-package.R
\docType{package}
\name{handwriter-package}
\alias{handwriter}
\alias{handwriter-package}
\title{handwriter: Handwriting Analysis in R}
\description{
Process handwriting document into letters, words, and lines. Provides measurements at all levels. Webpage provided at: \url{https://csafe-isu.github.io/handwriter/index.html}.
}
\author{
\strong{Maintainer}: James Taylor \email{jamesetay1@gmail.com}

Authors:
\itemize{
  \item Nick Berry \email{berryni@iastate.edu}
  \item Felix Baez-Santiago \email{felbasa@iastate.edu}
}

}
\keyword{internal}
