% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{i_to_rc}
\alias{i_to_rc}
\title{i_to_rc}
\usage{
i_to_rc(nodes, dims)
}
\arguments{
\item{nodes}{Nodes to be converted.}

\item{img_dim}{Dimensions of binary image}
}
\value{
Returns matrix mapping nodes to respective row, col.
}
\description{
Convert indicies to respective row, col.
}
\keyword{binary,}
\keyword{column,}
\keyword{image}
\keyword{row,}
