% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{letterPaths}
\alias{letterPaths}
\title{Internal function that uses existing breakPoint list to assign letters to the nodes in nodeGraph0.}
\usage{
letterPaths(allPaths, nodeGraph0, breakPoints)
}
\description{
Internal function that uses existing breakPoint list to assign letters to the nodes in nodeGraph0.
}
