% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFeatures.R
\name{line_number_extract}
\alias{line_number_extract}
\title{line_number_extract}
\usage{
line_number_extract(all_centroids, img_dim)
}
\arguments{
\item{all_centroids}{List of centroids extracted from cumulative character_features}

\item{img_dim}{Dimensions of binary image}
}
\value{
List associating line numbers to characters
}
\description{
Primary logic unit for line number to character association
}
\keyword{character,}
\keyword{features,}
\keyword{line,}
\keyword{number}
