% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london}
\alias{london}
\title{Cursive written word: London}
\format{Binary image matrix. 148 rows and 481 columns.}
\usage{
london
}
\description{
Cursive written word: London
}
\keyword{datasets}
