% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LetterListFunctions.R
\name{loop_extract}
\alias{loop_extract}
\title{loop_extract}
\usage{
loop_extract(allPaths)
}
\arguments{
\item{allPaths}{All character (formerly letter) paths from processHandwriting()}
}
\description{
Iterates through all available paths from processHandwriting()
Picks out loops for later character association.
}
