% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusteringTemplates.R
\name{make_clustering_templates}
\alias{make_clustering_templates}
\title{make_clustering_templates}
\usage{
make_clustering_templates(
  template_dir,
  max_edges = 30,
  starting_seed = 100,
  K = 40,
  num_runs = 1,
  num_cores = 1,
  num_dist_cores = 1,
  num_path_cuts = 8,
  max_iters = 1,
  gamma = 3,
  num_graphs = "All"
)
}
\description{
TODO
}
