% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMakeTemplates.R,
%   R/MakeClusteringTemplates.R
\name{make_clustering_templates}
\alias{make_clustering_templates}
\title{make_clustering_templates}
\usage{
make_clustering_templates(
  template_dir,
  max_edges = 30,
  starting_seed = 100,
  K = 40,
  num_runs = 1,
  num_cores = 1,
  num_dist_cores = 1,
  num_path_cuts = 8,
  max_iters = 1,
  gamma = 3,
  num_graphs = "All"
)

make_clustering_templates(
  template_dir,
  max_edges = 30,
  starting_seed = 100,
  K = 40,
  num_runs = 1,
  num_cores = 1,
  num_dist_cores = 1,
  num_path_cuts = 8,
  max_iters = 1,
  gamma = 3,
  num_graphs = "All"
)
}
\arguments{
\item{template_dir}{Input directory}

\item{max_edges}{Maximum number of edges allowed in input graphs. Graphs with
more than the maximum number will be ignored.}

\item{starting_seed}{Integer seed for the random number generator. If
\code{num_runs} is 1 then a single cluster template is created with the starting
seed. If \code{num_runs} is greater than 1, multiple cluster templates will be
created by adding 1, 2, 3, and so on to the starting seed.}

\item{K}{Integer number of clusters}

\item{num_runs}{Integer number of cluster templates to create}

\item{num_cores}{Integer number of cores. If \code{num_runs} is greater than 1,
cluster templates will be created on different cores.}

\item{num_dist_cores}{Integer number of cores to use for the distance
calculations in the K-means algorithm. Each iteration of the K-means
algorithm calculates the distance between each input graph and each cluster
center.}

\item{num_path_cuts}{Integer number of sections to cut each graph into for
shape comparison}

\item{max_iters}{Maximum number of iterations to allow the K-means algorithm
to run}

\item{gamma}{Parameter for outliers}

\item{num_graphs}{Number of graphs to use to create the cluster template.
\code{All} uses all available graphs. An integer uses a random sample of graphs.}
}
\value{
List containing the cluster template(s)
}
\description{
\code{make_clustering_templates()} applies a K-means clustering algorithm to the
input handwriting samples pre-processed with
\code{\link[=process_batch_dir]{process_batch_dir()}} and saved in the input folder \verb{template_dir > data > template_graphs}. The K-means algorithm sorts the graphs in the input
handwriting samples into groups, or clusters, of similar graphs.

TODO
}
\keyword{?}
