% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{message}
\alias{message}
\title{Full page image of the handwritten London letter.}
\format{
Binary image matrix. 1262 rows and 1162 columns.
}
\usage{
message
}
\description{
Full page image of the handwritten London letter.
}
\examples{
\dontrun{
message_document <- list()
message_document$image <- message
plotImage(message_document)
message_document$thin <- thinImage(message_document$image)
plotImageThinned(message_document)
message_processList <- processHandwriting(message_document$thin, dim(message_document$image))
}
}
\keyword{datasets}
