% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_wrapped_cauchy}
\alias{model_wrapped_cauchy}
\title{Wrapped-Cauchy RJAGS Model}
\format{
An RJAGS model written as text.
}
\usage{
model_wrapped_cauchy
}
\description{
The hierarchical model used in \code{\link[=fit_model]{fit_model()}} and described in Crawford, Amy
M., Danica M. Ommen, and Alicia L. Carriquiry. "A rotation-based feature and
Bayesian hierarchical model for the forensic evaluation of handwriting
evidence in a closed set." The Annals of Applied Statistics 17, no. 2 (2023):
1127-1151.
}
\examples{
# view the Wrapped-Cauchy RJAGS model
model_wrapped_cauchy

}
\keyword{datasets}
