% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{pathLetterAssociate}
\alias{pathLetterAssociate}
\title{Function associating entries in allPaths to each letter}
\usage{
pathLetterAssociate(allPaths, letter)
}
\description{
Function associating entries in allPaths to each letter
}
