% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_cluster_fill_counts}
\alias{plot_cluster_fill_counts}
\title{Plot Cluster Fill Counts}
\usage{
plot_cluster_fill_counts(formatted_data, writerID, docname)
}
\arguments{
\item{formatted_data}{A list of formatted data created by
\code{format_model_data()} or \code{format_questioned_data}.}

\item{writerID}{An integer writer ID}

\item{doc}{A document name as a character string}
}
\value{
ggplot barchart of cluster fill counts
}
\description{
Create a barchart of the cluster fill counts for a particular writer and
document whose data is stored in formatted data.
}
\examples{
plot_cluster_fill_counts(formatted_data = example_model_data, 
                         writerID = 9, 
                         docname = "s01_pWOZ_r1.")

}
