% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_posterior_probabilities}
\alias{plot_posterior_probabilities}
\title{plot_posterior_probabilities}
\usage{
plot_posterior_probabilities(analysis)
}
\arguments{
\item{analysis}{A named list of analysis results from \code{\link[=analyze_questioned_documents]{analyze_questioned_documents()}}.}
}
\value{
A tile plot of posterior probabilities of writership.
}
\description{
Creates a tile plot of posterior probabilities of writership for each
questioned document and each known writer analyzed with
\code{\link[=analyze_questioned_documents]{analyze_questioned_documents()}}.
}
\examples{
\dontrun{
draws <- fit_model(example_model_training_data, num_iters = 4000)
draws <- drop_burnin(draws, 1000)
analysis <- analyze_questioned_documents(example_model_training_data, draws, example_questioned_data, num_cores = 4)
}

}
