% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_posterior_probabilities}
\alias{plot_posterior_probabilities}
\title{Plot Posterior Probabilities}
\usage{
plot_posterior_probabilities(analysis)
}
\arguments{
\item{analysis}{A named list of analysis results from \code{\link[=analyze_questioned_documents]{analyze_questioned_documents()}}.}
}
\value{
A tile plot of posterior probabilities of writership.
}
\description{
Creates a tile plot of posterior probabilities of writership for each
questioned document and each known writer analyzed with
\code{\link[=analyze_questioned_documents]{analyze_questioned_documents()}}.
}
\examples{
plot_posterior_probabilities(analysis = example_analysis)

}
