% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace}
\usage{
plot_trace(model, variable)
}
\arguments{
\item{model}{An MCMC object created by \code{\link{fit_model}}}

\item{variable}{The name of a variable in the MCMC object}
}
\value{
ggplot line plot
}
\description{
Create a traceplot for all chains for a single variable in an MCMC object
created by \code{\link{fit_model}}. If the model contains more than one chain, the
chains will be combined by pasting them together.
}
\examples{
model <- fit_model(example_model_data, num_iters = 500, num_chains = 1)
plot_trace(model = model, variable = "theta[1,1]")

}
