% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace}
\usage{
plot_trace(variable, model)
}
\arguments{
\item{variable}{The name of a variable in the model}

\item{model}{A model created by \code{\link{fit_model}}}
}
\value{
A trace plot
}
\description{
Create a traceplot for all chains for a single variable in an MCMC object
created by \code{\link{fit_model}}. If the model contains more than one chain, the
chains will be combined by pasting them together.
}
\examples{
plot_trace(model = example_model_1chain, variable = "pi[1,1]")
plot_trace(model = example_model_1chain, variable = "mu[2,3]")
plot_trace(model = example_model_1chain, variable = "gamma[1]")
plot_trace(model = example_model_1chain, variable = "tau[2,3]")
plot_trace(model = example_model_1chain, variable = "eta[2]")

}
