% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterModeling_plottingfunctions.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace}
\usage{
plot_trace(model, model_data, variable)
}
\arguments{
\item{model}{An MCMC object created by \code{\link{fit_model}}}

\item{model_data}{Data output by \code{\link{format_model_data}} and input to \code{\link{fit_model}}}

\item{variable}{The name of a variable in the MCMC object}
}
\value{
ggplot line plot
}
\description{
Create a traceplot for all chains for a single variable in an MCMC object
created by \code{\link{fit_model}}. If the model contains more than one chain, the
chains will be combined by pasting them together.
}
\examples{
model <- fit_model(model_data = example_model_data, num_iters = 500, num_chains = 1)
plot_trace(model = model, model_data = example_model_data, variable = "pi[1,1]")
plot_trace(model = model, model_data = example_model_data, variable = "mu[2,3]")
plot_trace(model = model, model_data = example_model_data, variable = "gamma[1]")
plot_trace(model = model, model_data = example_model_data, variable = "tau[2,3]")
plot_trace(model = model, model_data = example_model_data, variable = "eta[2]")

}
