% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriangulationPlotFunctions.R
\name{plotColorNodes}
\alias{plotColorNodes}
\title{plotColorNodes}
\usage{
plotColorNodes(letterList, whichWord, dims, wordInfo)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{whichWord}{Single word value denoting which line to plot - checked if too big inside function.}

\item{dims}{Dimensions of the original document}
}
\value{
Plot of single word.
}
\description{
This function returns a plot of a single Word extracted from a document.
It plots the color as well
}
