% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotImage}
\alias{plotImage}
\title{Plot Image}
\usage{
plotImage(doc)
}
\arguments{
\item{doc}{A document processed with \code{\link[=processDocument]{processDocument()}} or a binary matrix (all entries are 0 or 1)}
}
\value{
ggplot plot
}
\description{
This function plots a basic black and white image.
}
\examples{
csafe_document <- list()
csafe_document$image <- csafe
plotImage(csafe_document)

\dontrun{
document <- processDocument('path/to/image.png')
plotImage(document)
}

}
