% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotImage}
\alias{plotImage}
\title{plotImage}
\usage{
plotImage(x)
}
\arguments{
\item{x}{Binary matrix, usually from readPNGBinary}
}
\value{
Returns plot of x.
}
\description{
This function plots a basic binary image.
}
\examples{
nature1_document = list()
nature1_document$image = nature1
plotImage(nature1_document$image)

}
\keyword{plot}
