% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{plotImage}
\alias{plotImage}
\title{plotImage}
\usage{
plotImage(x)
}
\arguments{
\item{x}{Binary matrix, usually from readPNGBinary}
}
\value{
Returns plot of x.
}
\description{
This function plots a basic binary image.
}
\keyword{plot}
