% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotImageThinned}
\alias{plotImageThinned}
\title{Plot Thinned Image}
\usage{
plotImageThinned(doc)
}
\arguments{
\item{doc}{A document processed with \code{\link[=processHandwriting]{processHandwriting()}}}
}
\value{
gpplot plot of thinned image
}
\description{
This function returns a plot with the full image plotted in light gray and the thinned skeleton printed in black on top.
}
\examples{
csafe_document <- list()
csafe_document$image <- csafe
csafe_document$thin <- thinImage(csafe_document$image)
plotImageThinned(csafe_document)

}
