% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotImageThinned}
\alias{plotImageThinned}
\title{plotImageThinned}
\usage{
plotImageThinned(doc)
}
\arguments{
\item{doc}{A document processed with [handwriter::processHandwriting()]}
}
\value{
Plot of thinned image
}
\description{
This function returns a plot with the full image plotted in light gray and the skeleton printed in black on top.
}
\examples{
\dontrun{
document = processHandwriting('path/to/image.png')
plotImageThinned(document)
}
}
