% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotImageThinned}
\alias{plotImageThinned}
\title{plotImageThinned}
\usage{
plotImageThinned(img, thinned)
}
\arguments{
\item{img}{Full image matrix}

\item{thinned}{Thinned image matrix}
}
\value{
Plot of full and thinned image.
}
\description{
This function returns a plot with the full image plotted in light gray and the skeleton printed in black on top.
}
\examples{

## Not Run
# plotImageThinned(london, london_thin)
# plotImageThinned(message, message_thin)

}
