% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotLetter}
\alias{plotLetter}
\title{plotLetter}
\usage{
plotLetter(letterList, whichLetter, dims, showPaths = TRUE,
  showCentroid = TRUE, showSlope = TRUE)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{whichLetter}{Single value in 1:length(letterList) denoting which letter to plot.}

\item{dims}{Dimensions of the original document}

\item{showPaths}{Whether the calculated paths on the letter should be shown with numbers.}
}
\value{
Plot of single letter.
}
\description{
This function returns a plot of a single letter extracted from a document. It uses the letterList parameter from the processHandwriting function and accepts a single value as whichLetter. Dims requires the dimensions of the entire document, since this isn't contained in processHandwriting.
}
