% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotLine}
\alias{plotLine}
\title{plotLine}
\usage{
plotLine(letterList, whichLine, dims)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{whichLine}{Single value denoting which line to plot - checked if too big inside function.}

\item{dims}{Dimensions of the original document}
}
\value{
Plot of single line.
}
\description{
This function returns a plot of a single line extracted from a document. It uses the letterList parameter from the processHandwriting function and accepts a single value as whichLetter. Dims requires the dimensions of the entire document, since this isn't contained in processHandwriting.
}
