% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotNodes}
\alias{plotNodes}
\title{plotNodes}
\usage{
plotNodes(doc, plot_break_pts = FALSE, nodeSize = 3, nodeColor = "red")
}
\arguments{
\item{doc}{A document processed with [handwriter::processHandwriting()]}

\item{plot_break_pts}{Logical value as to whether to plot nodes or break points. plot_break_pts=FALSE plots nodes and plot_break_pts=TRUE plots break point.}

\item{nodeSize}{Size of triangles printed. 3 by default. Move down to 2 or 1 for small text images.}

\item{nodeColor}{Which color the nodes should be}
}
\value{
Plot of full and thinned image with vertices overlaid.
}
\description{
This function returns a plot with the full image plotted in light gray and the skeleton printed in black, with red triangles over the vertices.
Also called from plotPath, which is a more useful function, in general.
}
\examples{
\dontrun{
twoSent_doc = list()
twoSent_doc$image = twoSent
twoSent_doc$thin = thinImage(twoSent_doc$image)
twoSent_doc$process = processHandwriting(twoSent_doc$thin, dim(twoSent_doc$image))
plotNodes(twoSent_doc)
}
}
