% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{process_batch_dir}
\alias{process_batch_dir}
\title{process_batch_dir}
\usage{
process_batch_dir(input_dir, output_dir = ".", transform_output = "document")
}
\arguments{
\item{input_dir}{Input directory that contains images}

\item{output_dir}{A directory to save the processed images}

\item{transform_output}{The type of transformation to perform on the output}
}
\value{
A list of processed documents. The list is also saved to output_dir.
}
\description{
process_batch_dir
}
\keyword{?}
