% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{process_batch_dir}
\alias{process_batch_dir}
\title{Process Batch Directory}
\usage{
process_batch_dir(input_dir, output_dir = ".")
}
\arguments{
\item{input_dir}{Input directory that contains images}

\item{output_dir}{A directory to save the processed images}
}
\value{
A list of processed documents
}
\description{
Process a list of handwriting samples saved as PNG images in a directory:
(1) Load the image and convert it to black and white with \code{\link[=readPNGBinary]{readPNGBinary()}}
(2) Thin the handwriting to one pixel in width with \code{\link[=thinImage]{thinImage()}}
(3) Run \code{\link[=processHandwriting]{processHandwriting()}} to split the handwriting into parts called \emph{edges} and place \emph{nodes} at the ends of
edges. Then combine edges into component shapes called \emph{graphs}.
(4) Save the processed document in an RDS file.
(5) Optional. Return a list of the processed documents.
}
\examples{
\dontrun{
process_batch_list("path/to/input_dir", "path/to/output_dir")
docs <- process_batch_list("path/to/input_dir", "path/to/output_dir")
}

}
