% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{process_batch_dir}
\alias{process_batch_dir}
\title{process_batch_dir}
\usage{
process_batch_dir(
  image_batch,
  batch_output_dir = ".",
  transform_output = "none"
)
}
\arguments{
\item{batch_output_dir}{Output directory}

\item{batch_input_dir}{Input directory as a list of datapaths or a directory}
}
\description{
process_batch_dir
}
\keyword{?}
