% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{process_batch_list}
\alias{process_batch_list}
\title{process_batch_list}
\usage{
process_batch_list(image_batch, name_list = NULL, transform_output = "none")
}
\arguments{
\item{batch_input_dir}{Input directory as a list from shiny}
}
\description{
process_batch_list
}
\keyword{?}
