% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{process_batch_list}
\alias{process_batch_list}
\title{process_batch_list}
\usage{
process_batch_list(
  image_list,
  output_dir,
  return_result = TRUE,
  transform_output = "document"
)
}
\arguments{
\item{image_list}{A list of image file paths}

\item{output_dir}{A directory to save the processed images}

\item{return_result}{TRUE/FALSE whether to return the result. If TRUE, the
processed documents with be saved and a list of the processed documents
will be returned. If FALSE, the processed documents will be saved, but
nothing will be returned.}

\item{transform_output}{The type of transformation to perform on the output}
}
\value{
A list of processed documents. The list is also saved to output_dir.
}
\description{
process_batch_list
}
