% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordSeparation_old.R
\name{process_words}
\alias{process_words}
\title{process_words}
\usage{
process_words(words, dims, triangulate = FALSE)
}
\arguments{
\item{words}{List of words and some glyph level information}

\item{dims}{The dimensions of the image (important for r/c features)}

\item{triangulate}{Logical value that begins the triangulation process when set to TRUE.}
}
\value{
A new list with word level information for each word.
}
\description{
Gets information on a word level
}
\examples{
twoSent_document = list()
twoSent_document$image = twoSent
twoSent_document$thin = thinImage(twoSent_document$image)
twoSent_processList = processHandwriting(twoSent_document$thin, dim(twoSent_document$image))

dims = dim(twoSent_document$image)
words = create_words(twoSent_processList) 
words_after_processing = process_words(words, dim(twoSent_document$image), TRUE)

}
