% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractWordInfo.R
\name{process_words}
\alias{process_words}
\title{process_words}
\usage{
process_words(words, dims)
}
\arguments{
\item{words}{List of words and some glyph level information}

\item{dims}{The dimensions of the image (important for r/c features)}
}
\value{
A new list with word level information for each word.
}
\description{
gets information on a word level
}
