% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertemplate_docprocessing.R
\name{processDocsForClustering}
\alias{processDocsForClustering}
\title{processDocsForClustering}
\usage{
processDocsForClustering(documentDirectory, numCores)
}
\description{
Some stuff here
}
