% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{processDocument}
\alias{processDocument}
\title{processDocument}
\usage{
processDocument(path)
}
\arguments{
\item{path}{File path for handwriting document. The document must be in PNG file format.}
}
\value{
A named list
}
\description{
Load a handwriting sample from a PNG image. Then binarize, thin, and split the handwriting into graphs.
}
