% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{read_and_process}
\alias{read_and_process}
\title{Read and Process}
\usage{
read_and_process(image_name, transform_output)
}
\arguments{
\item{image_name}{The file path to an image}

\item{transform_output}{The type of transformation to perform on the output}
}
\value{
A list of the processed image components
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Development on \code{read_and_process()} is complete. We recommend using \code{\link[=processDocument]{processDocument()}}.
\code{read_and_process(image_name, "document")} is equivalent to \code{processDocument(image_name)}.
}
\examples{
# use handwriting example from handwriter package
image_path <- system.file("extdata", "phrase_example.png", package = "handwriter")
doc <- read_and_process(image_path, "document")

}
