% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchProcessing.R
\name{read_and_process}
\alias{read_and_process}
\title{read_and_process}
\usage{
read_and_process(image_name, transform_output)
}
\arguments{
\item{image_name}{The file path to an image}

\item{transform_output}{The type of transformation to perform on the output}
}
\value{
A list of the processed image components
}
\description{
read_and_process
}
