% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{readPNGBinary}
\alias{readPNGBinary}
\title{readPNGBinary}
\usage{
readPNGBinary(path, binaryCutoff = "auto", inversion = FALSE)
}
\arguments{
\item{path}{File path for image.}

\item{binaryCutoff}{Value to cut off non binary images. Leave at "auto" to let k-means pick.
Put in "xx\%" to set default threshold. Enter numeric value to scale auto threshold by that amount.
binaryCutoff is passed to imager::threshold.}

\item{inversion}{Boolean dictating whether or not to flip each pixel of binarized image. Flipping happens after binarization. FALSE by default.}
}
\value{
Returns image from path. 0 represents black, and 1 represents white by default.
}
\description{
This function reads in and binarizes PNG images from the specified file path.
}
\keyword{binary}
