% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{readPNGBinary}
\alias{readPNGBinary}
\title{readPNGBinary}
\usage{
readPNGBinary(path, binarizationCutoff = 0, inversion = FALSE)
}
\arguments{
\item{path}{File path for image.}

\item{binarizationCutoff}{Value to cut off non binary images. Values higher than this value go to 1 and lower than this value go to 0. 0 by default, no binarization.}

\item{inversion}{Boolean dictating whether or not to flip each bin of binarized image. Flipping happens after binarization. FALSE by default.}
}
\value{
Returns image from path. 0 represents black, and 1 represents white by default.
}
\description{
This function reads in and binarizes PNG images from the specified file path.
}
\keyword{PNG,}
\keyword{binary}
