% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{readPNGBinary}
\alias{readPNGBinary}
\title{readPNGBinary
This function reads in and binarizes PNG images from the specified file path.}
\usage{
readPNGBinary(path, cutoffAdjust = 1, clean = TRUE, crop = TRUE,
  inversion = FALSE)
}
\arguments{
\item{path}{File path for image.}

\item{cutoffAdjust}{Multiplicative adjustment to the K-means estimated binarization cutoff.}

\item{clean}{Whether to fill in white pixels with 7 or 8 neighbors. This will help a lot when thinning -- keeps from getting little white bubbles in text.}

\item{inversion}{Boolean dictating whether or not to flip each pixel of binarized image. Flipping happens after binarization. FALSE by default.}
}
\value{
Returns image from path. 0 represents black, and 1 represents white by default.
}
\description{
readPNGBinary
This function reads in and binarizes PNG images from the specified file path.
}
\keyword{binary}
