% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rgb2grayscale}
\alias{rgb2grayscale}
\title{rgba2grayscale}
\usage{
rgb2grayscale(img)
}
\arguments{
\item{img}{A 3D array with slices R, G, and B}
}
\value{
img as a 3D array as grayscale
}
\description{
Changes RGB image to grayscale
}
