% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThinText.R
\name{rgba2rgb}
\alias{rgba2rgb}
\title{rgba2rgb}
\usage{
rgba2rgb(rgba)
}
\arguments{
\item{rgba}{3-d vector with 4 channels. [n, p, # channels]}
}
\value{
3-d vector with 3 channels [n, p, 3]
}
\description{
Internal function for converting rgb with alpha channel to standard rgb.
}
