% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeansFunctions.R
\name{runAndSaveKmeans}
\alias{runAndSaveKmeans}
\title{runAndSaveKmeans}
\usage{
runAndSaveKmeans(
  imList,
  centers,
  numDistCores = 0,
  K,
  iter_number,
  numPathCuts = 8,
  iter.max = 500,
  numOutliers = round(0.25 * length(imList)),
  gamma = 3,
  results_file,
  outfile,
  meanGraph = "slow_change",
  meanGraphOrder = "sequential",
  fillClusters = "farthest"
)
}
\arguments{
\item{imList}{List graphs from handwriting samples}

\item{centers}{List of starting cluster centers}

\item{numDistCores}{Integer number of cores to use for distance calculations}

\item{K}{Integer number of clusters. If numOutliers > 0, then
there will be K+1 clusters.}

\item{iter_number}{Integer iteration number. This will be used in the file name of the 
saved clustering template.}

\item{numPathCuts}{Integer number of segments to cut each path into}

\item{iter.max}{Integer maximum number of iterations to allow cluster centers to converge}

\item{numOutliers}{Integer number of outliers to allow. If numOutliers is zero, no outliers will be
assigned.}

\item{gamma}{Float parameter for calculating the outlier cutoff. 
If numOutliers is zero, gamma has no effect.}

\item{outfile}{String file path and name of log file}

\item{meanGraph}{String choice of algorithm for calculating mean graphs. The choices are 'basic',
'slow_change' and 'kmeans'}

\item{meanGraphOrder}{Chose the order in which to add graphs to the mean 
graph calculations. The options are 'sequential' and 'random'}

\item{fillClusters}{String choice of how to choose which graph to reassign to an empty cluster. 
Choice are 'random' and 'farthest'}

\item{numberToRun}{Integer number of clustering templates to create}
}
\value{
Clustering template
}
\description{
Create and save a signle clustering template.
}
\keyword{?}
