% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{SaveAllLetterPlots}
\alias{SaveAllLetterPlots}
\title{SaveAllLetterPlots}
\usage{
SaveAllLetterPlots(letterList, filePaths, documentDimensions,
  bgTransparent = TRUE)
}
\arguments{
\item{letterList}{Letter list from processHandwriting function}

\item{filePaths}{Folder path to save images to}

\item{documentDimensions}{Dimensions of original document}
}
\value{
Nothing
}
\description{
This function returns a plot of a single letter extracted from a document. It uses the letterList parameter from the processHandwriting function and accepts a single value as whichLetter. Dims requires the dimensions of the entire document, since this isn't contained in processHandwriting. Requires the \pkg{\link{magick}} package.
}
\seealso{
\pkg{\link{magick}}
}
