% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDistanceFunctions.R
\name{solveLP}
\alias{solveLP}
\title{solveLP}
\usage{
solveLP(dists)
}
\arguments{
\item{dists}{?}
}
\value{
A list of information about the optimal pairings
}
\description{
Use linear programming to solve a constrained minimization optimization
problem to find the optimal pairings of paths between two graphs.
}
\keyword{?}
