% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thinImage}
\alias{thinImage}
\title{thinImage}
\usage{
thinImage(img)
}
\arguments{
\item{img}{A binary matrix of the text that is to be thinned.}
}
\value{
A thinned, one pixel wide, image.
}
\description{
This function returns a vector of locations for black pixels in the thinned image.
Thinning done using Zhang - Suen algorithm.
}
