% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JunctionDetection.R
\name{whichNeighbors}
\alias{whichNeighbors}
\title{whichNeighbors}
\usage{
whichNeighbors(coords, img)
}
\arguments{
\item{coords}{coordinates to consider}

\item{img}{The image as a bitmap}
}
\value{
Return a matrix of which neighbors are a black pixel
}
\description{
Internal function for identifying which neighbors are black.
}
